Move metafunction work into reflect.h2
Instead of a separate file in ./source
