+++ regression-tests/test-results/msvc-2022-c++latest/pure2-regex_10_escapes.cpp.execution
