+++ source/reflect.h
@@ -10,36 +10,115 @@
 
 #line 1 "reflect.h2"
 
-#line 22 "reflect.h2"
+#line 21 "reflect.h2"
 namespace cpp2 {
 
 namespace meta {
 
-#line 34 "reflect.h2"
+#line 33 "reflect.h2"
 class compiler_services;
 
-#line 228 "reflect.h2"
+#line 227 "reflect.h2"
 class declaration_base;
 
-#line 254 "reflect.h2"
+#line 253 "reflect.h2"
 class declaration;
 
-#line 336 "reflect.h2"
+#line 335 "reflect.h2"
 class function_declaration;
 
-#line 426 "reflect.h2"
+#line 425 "reflect.h2"
 class object_declaration;
 
-#line 462 "reflect.h2"
+#line 461 "reflect.h2"
 class type_declaration;
 
-#line 599 "reflect.h2"
+#line 598 "reflect.h2"
 class alias_declaration;
 
-#line 1007 "reflect.h2"
+#line 1006 "reflect.h2"
 class value_member_info;
 
-#line 1693 "reflect.h2"
+#line 1505 "reflect.h2"
+}
+
+#line 1522 "reflect.h2"
+namespace regex {
+
+#line 1528 "reflect.h2"
+class expression_flags;
+
+#line 1544 "reflect.h2"
+class regex_token;
+
+#line 1570 "reflect.h2"
+class regex_token_check;
+
+#line 1589 "reflect.h2"
+class regex_token_code;
+
+#line 1608 "reflect.h2"
+class regex_token_empty;
+
+#line 1624 "reflect.h2"
+class regex_token_list;
+
+#line 1663 "reflect.h2"
+class parse_context_group_state;
+
+#line 1724 "reflect.h2"
+class parse_context_branch_reset_state;
+
+#line 1767 "reflect.h2"
+class parse_context;
+
+#line 2165 "reflect.h2"
+class generation_function_context;
+    
+
+#line 2183 "reflect.h2"
+class generation_context;
+
+#line 2381 "reflect.h2"
+class alternative_token;
+
+#line 2396 "reflect.h2"
+class alternative_token_gen;
+
+#line 2448 "reflect.h2"
+class any_token;
+
+#line 2466 "reflect.h2"
+class char_token;
+
+#line 2569 "reflect.h2"
+class class_token;
+
+#line 2784 "reflect.h2"
+class group_ref_token;
+
+#line 2915 "reflect.h2"
+class group_token;
+
+#line 3202 "reflect.h2"
+class lookahead_token;
+
+#line 3283 "reflect.h2"
+class range_token;
+
+#line 3431 "reflect.h2"
+class special_range_token;
+
+#line 3498 "reflect.h2"
+template<typename Error_out> class regex_generator;
+
+#line 3572 "reflect.h2"
+}
+
+#line 3575 "reflect.h2"
+namespace meta {
+
+#line 3749 "reflect.h2"
 }
 
 }
@@ -67,25 +146,17 @@ class value_member_info;
 
 #include "parse.h"
 #include "cpp2regex.h"
-#include "regex.h"
 
-#line 22 "reflect.h2"
+#line 21 "reflect.h2"
 namespace cpp2 {
 
 namespace meta {
 
-#line 27 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//  Compiler services
-//
-//-----------------------------------------------------------------------
-//
-
+#line 33 "reflect.h2"
 class compiler_services
  {
-    //  Common data members
-    //
+
+#line 37 "reflect.h2"
     private: std::vector<error_entry>* errors; 
     private: int errors_original_size; 
     private: stable_vector<token>* generated_tokens; 
@@ -94,119 +165,64 @@ class compiler_services
     private: std::vector<std::string> metafunction_args {}; 
     private: bool metafunctions_used {false}; 
 
-    //  Constructor
-    //
+#line 47 "reflect.h2"
     public: explicit compiler_services(
 
         std::vector<error_entry>* errors_, 
         stable_vector<token>* generated_tokens_
     );
 
-#line 60 "reflect.h2"
-    //  Common API
-    //
+#line 61 "reflect.h2"
     public: auto set_metafunction_name(cpp2::impl::in<std::string_view> name, cpp2::impl::in<std::vector<std::string>> args) & -> void;
 
-#line 68 "reflect.h2"
+#line 67 "reflect.h2"
     public: [[nodiscard]] auto get_metafunction_name() const& -> std::string_view;
 
     public: [[nodiscard]] auto get_argument(cpp2::impl::in<int> index) & -> std::string;
 
-#line 78 "reflect.h2"
+#line 77 "reflect.h2"
     public: [[nodiscard]] auto get_arguments() & -> std::vector<std::string>;
 
-#line 83 "reflect.h2"
+#line 82 "reflect.h2"
     public: [[nodiscard]] auto arguments_were_used() const& -> bool;
 using parse_statement_ret = std::unique_ptr<statement_node>;
 
 
-#line 85 "reflect.h2"
+#line 84 "reflect.h2"
     protected: [[nodiscard]] auto parse_statement(
 
         std::string_view source
     ) & -> parse_statement_ret;
 
-#line 138 "reflect.h2"
+#line 137 "reflect.h2"
     public: [[nodiscard]] virtual auto position() const -> source_position;
 
-#line 144 "reflect.h2"
-    //  Error diagnosis and handling, integrated with compiler output
-    //  Unlike a contract violation, .requires continues further processing
-    //
+#line 146 "reflect.h2"
     public: auto require(
 
         cpp2::impl::in<bool> b, 
         cpp2::impl::in<std::string_view> msg
     ) const& -> void;
 
-#line 158 "reflect.h2"
+#line 157 "reflect.h2"
     public: auto error(cpp2::impl::in<std::string_view> msg) const& -> void;
 
-#line 167 "reflect.h2"
-    //  Enable custom contracts on this object, integrated with compiler output
-    //  Unlike .requires, a contract violation stops further processing
-    //
+#line 169 "reflect.h2"
     public: auto report_violation(auto const& msg) const& -> void;
 
-#line 178 "reflect.h2"
+#line 177 "reflect.h2"
     public: [[nodiscard]] auto is_active() const& -> auto;
     public: virtual ~compiler_services() noexcept;
 public: compiler_services(compiler_services const& that);
 
-#line 179 "reflect.h2"
+#line 178 "reflect.h2"
 };
 
-#line 182 "reflect.h2"
-/*
-//-----------------------------------------------------------------------
-//
-//  Type IDs
-//
-//-----------------------------------------------------------------------
-//
-
-//  All type_ids are wrappers around a pointer to node
-//
-type_id: @polymorphic_base @copyable type =
-{
-    this: compiler_services = ();
-
-    n: type_id_node;
-
-    protected operator=: (
-        out this,
-        n_: type_id_node,
-        s : compiler_services
-    )
-    = {
-        compiler_services = s;
-        n = n_;
-        assert( n, "a meta::type_id must point to a valid type_id_node, not null" );
-    }
-
-    is_wildcard         : (this) -> bool        = n.is_wildcard();
-    is_pointer_qualified: (this) -> bool        = n.is_pointer_qualified();
-    template_args_count : (this) -> int         = n.template_arguments().ssize();
-    to_string           : (this) -> std::string = n.to_string();
-
-    position: (override this) -> source_position = n.position();
-}
-*/
-
-#line 219 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//  Declarations
-//
-//-----------------------------------------------------------------------
-//
-
-//  All declarations are wrappers around a pointer to node
-//
+#line 227 "reflect.h2"
 class declaration_base
 : public compiler_services {
 
-#line 232 "reflect.h2"
+#line 231 "reflect.h2"
     protected: declaration_node* n; 
 
     protected: explicit declaration_base(
@@ -215,31 +231,28 @@ class declaration_base
         cpp2::impl::in<compiler_services> s
     );
 
-#line 245 "reflect.h2"
+#line 244 "reflect.h2"
     public: [[nodiscard]] auto position() const -> source_position override;
 
     public: [[nodiscard]] auto print() const& -> std::string;
     public: virtual ~declaration_base() noexcept;
 public: declaration_base(declaration_base const& that);
 
-#line 248 "reflect.h2"
+#line 247 "reflect.h2"
 };
 
-#line 251 "reflect.h2"
-//-----------------------------------------------------------------------
-//  All declarations
-//
+#line 253 "reflect.h2"
 class declaration
 : public declaration_base {
 
-#line 258 "reflect.h2"
+#line 257 "reflect.h2"
     public: explicit declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     );
 
-#line 267 "reflect.h2"
+#line 266 "reflect.h2"
     public: [[nodiscard]] auto is_public() const& -> bool;
     public: [[nodiscard]] auto is_protected() const& -> bool;
     public: [[nodiscard]] auto is_private() const& -> bool;
@@ -258,7 +271,7 @@ class declaration
 
     public: [[nodiscard]] auto name() const& -> std::string_view;
 
-#line 288 "reflect.h2"
+#line 287 "reflect.h2"
     public: [[nodiscard]] auto has_initializer() const& -> bool;
 
     public: [[nodiscard]] auto is_global() const& -> bool;
@@ -299,26 +312,22 @@ class declaration
     public: virtual ~declaration() noexcept;
 public: declaration(declaration const& that);
 
-                                                    // this precondition should be sufficient ...
 
-#line 330 "reflect.h2"
+#line 329 "reflect.h2"
 };
 
-#line 333 "reflect.h2"
-//-----------------------------------------------------------------------
-//  Function declarations
-//
+#line 335 "reflect.h2"
 class function_declaration
 : public declaration {
 
-#line 340 "reflect.h2"
+#line 339 "reflect.h2"
     public: explicit function_declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     );
 
-#line 350 "reflect.h2"
+#line 349 "reflect.h2"
     public: [[nodiscard]] auto index_of_parameter_named(cpp2::impl::in<std::string_view> s) const& -> int;
     public: [[nodiscard]] auto has_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool;
     public: [[nodiscard]] auto has_in_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool;
@@ -358,7 +367,7 @@ class function_declaration
 
     public: [[nodiscard]] auto get_parameters() const& -> std::vector<object_declaration>;
 
-#line 397 "reflect.h2"
+#line 396 "reflect.h2"
     public: [[nodiscard]] auto is_binary_comparison_function() const& -> bool;
 
     public: auto default_to_virtual() & -> void;
@@ -369,103 +378,94 @@ class function_declaration
     public: function_declaration(function_declaration const& that);
 
 
-#line 420 "reflect.h2"
+#line 419 "reflect.h2"
 };
 
-#line 423 "reflect.h2"
-//-----------------------------------------------------------------------
-//  Object declarations
-//
+#line 425 "reflect.h2"
 class object_declaration
 : public declaration {
 
-#line 430 "reflect.h2"
+#line 429 "reflect.h2"
     public: explicit object_declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     );
 
-#line 440 "reflect.h2"
+#line 439 "reflect.h2"
     public: [[nodiscard]] auto is_const() const& -> bool;
     public: [[nodiscard]] auto has_wildcard_type() const& -> bool;
 
     public: [[nodiscard]] auto type() const& -> std::string;
 
-#line 450 "reflect.h2"
+#line 449 "reflect.h2"
     public: [[nodiscard]] auto initializer() const& -> std::string;
     public: object_declaration(object_declaration const& that);
 
 
-#line 456 "reflect.h2"
+#line 455 "reflect.h2"
 };
 
-#line 459 "reflect.h2"
-//-----------------------------------------------------------------------
-//  Type declarations
-//
+#line 461 "reflect.h2"
 class type_declaration
 : public declaration {
 
-#line 466 "reflect.h2"
+#line 465 "reflect.h2"
     public: explicit type_declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     );
 
-#line 476 "reflect.h2"
+#line 475 "reflect.h2"
     public: auto reserve_names(cpp2::impl::in<std::string_view> name, auto&& ...etc) const& -> void;
 
-#line 490 "reflect.h2"
+#line 489 "reflect.h2"
     public: [[nodiscard]] auto is_polymorphic() const& -> bool;
     public: [[nodiscard]] auto is_final() const& -> bool;
     public: [[nodiscard]] auto make_final() & -> bool;
 
     public: [[nodiscard]] auto get_member_functions() const& -> std::vector<function_declaration>;
 
-#line 505 "reflect.h2"
+#line 504 "reflect.h2"
     public: [[nodiscard]] auto get_member_functions_needing_initializer() const& -> std::vector<function_declaration>;
 
-#line 520 "reflect.h2"
+#line 519 "reflect.h2"
     public: [[nodiscard]] auto get_member_objects() const& -> std::vector<object_declaration>;
 
-#line 530 "reflect.h2"
+#line 529 "reflect.h2"
     public: [[nodiscard]] auto get_member_types() const& -> std::vector<type_declaration>;
 
-#line 540 "reflect.h2"
+#line 539 "reflect.h2"
     public: [[nodiscard]] auto get_member_aliases() const& -> std::vector<alias_declaration>;
 
-#line 550 "reflect.h2"
+#line 549 "reflect.h2"
     public: [[nodiscard]] auto get_members() const& -> std::vector<declaration>;
 struct query_declared_value_set_functions_ret { bool out_this_in_that; bool out_this_move_that; bool inout_this_in_that; bool inout_this_move_that; };
 
 
 
-#line 560 "reflect.h2"
+#line 559 "reflect.h2"
     public: [[nodiscard]] auto query_declared_value_set_functions() const& -> query_declared_value_set_functions_ret;
 
-#line 575 "reflect.h2"
+#line 574 "reflect.h2"
     public: auto add_member(cpp2::impl::in<std::string_view> source) & -> void;
 
-#line 589 "reflect.h2"
+#line 588 "reflect.h2"
     public: auto remove_marked_members() & -> void;
     public: auto remove_all_members() & -> void;
 
     public: auto disable_member_function_generation() & -> void;
     public: type_declaration(type_declaration const& that);
 
-#line 593 "reflect.h2"
+#line 592 "reflect.h2"
 };
 
-#line 596 "reflect.h2"
-//-----------------------------------------------------------------------
-//  Alias declarations
-//
+#line 598 "reflect.h2"
 class alias_declaration
 : public declaration {
 
-#line 603 "reflect.h2"
+#line 602 "reflect.h2"
     public: explicit alias_declaration(
 
         declaration_node* n_, 
@@ -474,233 +474,62 @@ class alias_declaration
     public: alias_declaration(alias_declaration const& that);
 
 
-#line 612 "reflect.h2"
+#line 611 "reflect.h2"
 };
 
-#line 615 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//  Metafunctions - these are hardwired for now until we get to the
-//  step of writing a Cpp2 interpreter to run inside the compiler
-//
-//-----------------------------------------------------------------------
-//
-
-//-----------------------------------------------------------------------
-//  Some common metafunction helpers (metafunctions are just functions,
-//  so they can be factored as usual)
-//
+#line 626 "reflect.h2"
 auto add_virtual_destructor(meta::type_declaration& t) -> void;
 
-#line 633 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//      "... an abstract base class defines an interface ..."
-//
-//          -- Stroustrup (The Design and Evolution of C++, 12.3.1)
-//
-//-----------------------------------------------------------------------
-//
-//  interface
-//
-//  an abstract base class having only pure virtual functions
-//
+#line 644 "reflect.h2"
 auto interface(meta::type_declaration& t) -> void;
 
-#line 672 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     "C.35: A base class destructor should be either public and
-//      virtual, or protected and non-virtual."
-//
-//     "[C.43] ... a base class should not be copyable, and so does not
-//      necessarily need a default constructor."
-//
-//          -- Stroustrup, Sutter, et al. (C++ Core Guidelines)
-//
-//-----------------------------------------------------------------------
-//
-//  polymorphic_base
-//
-//  A pure polymorphic base type that is not copyable, and whose
-//  destructor is either public and virtual or protected and nonvirtual.
-//
-//  Unlike an interface, it can have nonpublic and nonvirtual functions.
-//
+#line 690 "reflect.h2"
 auto polymorphic_base(meta::type_declaration& t) -> void;
 
-#line 716 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     "... A totally ordered type ... requires operator<=> that
-//      returns std::strong_ordering. If the function is not
-//      user-written, a lexicographical memberwise implementation
-//      is generated by default..."
-//
-//          -- P0707R4, section 3
-//
-//      Note: This feature derived from Cpp2 was already adopted
-//            into Standard C++ via paper P0515, so most of the
-//            heavy lifting is done by the Cpp1 C++20/23 compiler,
-//            including the memberwise default semantics
-//            (In contrast, cppfront has to do the work itself for
-//            default memberwise semantics for operator= assignment
-//            as those aren't yet part of Standard C++)
-//
-//-----------------------------------------------------------------------
-//
-
+#line 735 "reflect.h2"
 auto ordered_impl(
     meta::type_declaration& t, 
-    cpp2::impl::in<std::string_view> ordering// must be "strong_ordering" etc.
+    cpp2::impl::in<std::string_view> ordering
 ) -> void;
 
-#line 760 "reflect.h2"
-//-----------------------------------------------------------------------
-//  ordered - a totally ordered type
-//
-//  Note: the ordering that should be encouraged as default gets the nice name
-//
+#line 764 "reflect.h2"
 auto ordered(meta::type_declaration& t) -> void;
 
-#line 770 "reflect.h2"
-//-----------------------------------------------------------------------
-//  weakly_ordered - a weakly ordered type
-//
+#line 772 "reflect.h2"
 auto weakly_ordered(meta::type_declaration& t) -> void;
 
-#line 778 "reflect.h2"
-//-----------------------------------------------------------------------
-//  partially_ordered - a partially ordered type
-//
+#line 780 "reflect.h2"
 auto partially_ordered(meta::type_declaration& t) -> void;
 
-#line 787 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     "A value is ... a regular type. It must have all public
-//      default construction, copy/move construction/assignment,
-//      and destruction, all of which are generated by default
-//      if not user-written; and it must not have any protected
-//      or virtual functions (including the destructor)."
-//
-//          -- P0707R4, section 3
-//
-//-----------------------------------------------------------------------
-//
-//  copyable
-//
-//  A type with (copy and move) x (construction and assignment)
-//
+#line 802 "reflect.h2"
 auto copyable(meta::type_declaration& t) -> void;
 
-#line 830 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//  basic_value
-//
-//  A regular type: copyable, plus has public default construction
-//  and no protected or virtual functions
-//
+#line 836 "reflect.h2"
 auto basic_value(meta::type_declaration& t) -> void;
 
-#line 855 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     "A 'value' is a totally ordered basic_value..."
-//
-//          -- P0707R4, section 3
-//
-//  value - a value type that is totally ordered
-//
-//  Note: the ordering that should be encouraged as default gets the nice name
-//
+#line 864 "reflect.h2"
 auto value(meta::type_declaration& t) -> void;
 
-#line 871 "reflect.h2"
+#line 870 "reflect.h2"
 auto weakly_ordered_value(meta::type_declaration& t) -> void;
 
-#line 877 "reflect.h2"
+#line 876 "reflect.h2"
 auto partially_ordered_value(meta::type_declaration& t) -> void;
 
-#line 884 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     C.20: If you can avoid defining default operations, do
-//
-//     ##### Reason
-//
-//     It's the simplest and gives the cleanest semantics.
-//
-//     ...
-//
-//     This is known as "the rule of zero".
-//
-//          -- C++ Core Guidelines
-//             C.20: If you can avoid defining any default operations, do
-//             <https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rc-zero>
-//
-//-----------------------------------------------------------------------
-//
-//  cpp1_rule_of_zero
-//
-//  a type without declared copy/move/destructor functions
-//
+#line 905 "reflect.h2"
 auto cpp1_rule_of_zero(meta::type_declaration& t) -> void;
 
-#line 918 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     "By definition, a `struct` is a `class` in which members
-//      are by default `public`; that is,
-//
-//      	struct s { ...
-//
-//      is simply shorthand for
-//
-//          class s { public: ...
-//
-//      ... Which style you use depends on circumstances and taste.
-//      I usually prefer to use `struct` for classes that have all
-//      data `public`."
-//
-//          -- Stroustrup (The C++ Programming Language, 3rd ed., p. 234)
-//
-//-----------------------------------------------------------------------
-//
-//  struct
-//
-//  a type with only public bases, objects, and functions,
-//  no virtual functions, and no user-defined constructors
-//  (i.e., no invariants) or assignment or destructors.
-//
+#line 942 "reflect.h2"
 auto cpp2_struct(meta::type_declaration& t) -> void;
 
-#line 990 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     "C enumerations constitute a curiously half-baked concept. ...
-//      the cleanest way out was to deem each enumeration a separate type."
-//
-//          -- Stroustrup (The Design and Evolution of C++, 11.7)
-//
-//     "An enumeration is a distinct type ... with named constants"
-//
-//          -- ISO C++ Standard
-//
-//-----------------------------------------------------------------------
-//
-//  basic_enum
-//
-//  a type together with named constants that are its possible values
-//
+#line 1006 "reflect.h2"
 class value_member_info {
     public: std::string name; 
     public: std::string type; 
     public: std::string value; 
     public: value_member_info(auto const& name_, auto const& type_, auto const& value_);
 
-#line 1011 "reflect.h2"
+#line 1010 "reflect.h2"
 };
 
 auto basic_enum(
@@ -709,83 +538,889 @@ auto basic_enum(
     cpp2::impl::in<bool> bitwise
     ) -> void;
 
-#line 1268 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//    "An enum[...] is a totally ordered value type that stores a
-//     value of its enumerators's type, and otherwise has only public
-//     member variables of its enumerator's type, all of which are
-//     naturally scoped because they are members of a type."
-//
-//          -- P0707R4, section 3
-//
+#line 1276 "reflect.h2"
 auto cpp2_enum(meta::type_declaration& t) -> void;
 
-#line 1294 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     "flag_enum expresses an enumeration that stores values
-//      corresponding to bitwise-or'd enumerators. The enumerators must
-//      be powers of two, and are automatically generated [...] A none
-//      value is provided [...] Operators | and & are provided to
-//      combine and extract values."
-//
-//          -- P0707R4, section 3
-//
+#line 1303 "reflect.h2"
 auto flag_enum(meta::type_declaration& t) -> void;
 
-#line 1326 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//     "As with void*, programmers should know that unions [...] are
-//      inherently dangerous, should be avoided wherever possible,
-//      and should be handled with special care when actually needed."
-//
-//          -- Stroustrup (The Design and Evolution of C++, 14.3.4.1)
-//
-//     "C++17 needs a type-safe union... The implications of the
-//      consensus `variant` design are well understood and have been
-//      explored over several LEWG discussions, over a thousand emails,
-//      a joint LEWG/EWG session, and not to mention 12 years of
-//      experience with Boost and other libraries."
-//
-//          -- Axel Naumann, in P0088 (wg21.link/p0088),
-//             the adopted proposal for C++17 std::variant
-//
-//-----------------------------------------------------------------------
-//
-//  union
-//
-//  a type that contains exactly one of a fixed set of values at a time
-//
-
+#line 1349 "reflect.h2"
 auto cpp2_union(meta::type_declaration& t) -> void;
 
-#line 1497 "reflect.h2"
-//-----------------------------------------------------------------------
-//
-//  print - output a pretty-printed visualization of t
-//
+#line 1500 "reflect.h2"
 auto print(cpp2::impl::in<meta::type_declaration> t) -> void;
 
-#line 1507 "reflect.h2"
-//-----------------------------------------------------------------------
+#line 1505 "reflect.h2"
+}
+
+#line 1522 "reflect.h2"
+namespace regex {
+
+using error_func = std::function<void(cpp2::impl::in<std::string> x)>;
+
+#line 1528 "reflect.h2"
+class expression_flags
+ {
+private: cpp2::u8 _value; private: constexpr expression_flags(cpp2::impl::in<cpp2::i64> _val);
+
+private: constexpr auto operator=(cpp2::impl::in<cpp2::i64> _val) -> expression_flags& ;
+public: constexpr auto operator|=(expression_flags const& that) & -> void;
+public: constexpr auto operator&=(expression_flags const& that) & -> void;
+public: constexpr auto operator^=(expression_flags const& that) & -> void;
+public: [[nodiscard]] constexpr auto operator|(expression_flags const& that) const& -> expression_flags;
+public: [[nodiscard]] constexpr auto operator&(expression_flags const& that) const& -> expression_flags;
+public: [[nodiscard]] constexpr auto operator^(expression_flags const& that) const& -> expression_flags;
+public: [[nodiscard]] constexpr auto has(expression_flags const& that) const& -> bool;
+public: constexpr auto set(expression_flags const& that) & -> void;
+public: constexpr auto clear(expression_flags const& that) & -> void;
+public: static const expression_flags case_insensitive;
+public: static const expression_flags multiple_lines;
+public: static const expression_flags single_line;
+public: static const expression_flags no_group_captures;
+public: static const expression_flags perl_code_syntax;
+public: static const expression_flags perl_code_syntax_in_classes;
+public: static const expression_flags none;
+public: [[nodiscard]] constexpr auto get_raw_value() const& -> cpp2::u8;
+public: constexpr explicit expression_flags();
+public: constexpr expression_flags(expression_flags const& that);
+public: constexpr auto operator=(expression_flags const& that) -> expression_flags& ;
+public: constexpr expression_flags(expression_flags&& that) noexcept;
+public: constexpr auto operator=(expression_flags&& that) noexcept -> expression_flags& ;
+public: [[nodiscard]] auto operator<=>(expression_flags const& that) const& -> std::strong_ordering = default;
+public: [[nodiscard]] auto to_string_impl(cpp2::impl::in<std::string_view> prefix, cpp2::impl::in<std::string_view> separator) const& -> std::string;
+public: [[nodiscard]] auto to_string() const& -> std::string;
+public: [[nodiscard]] auto to_code() const& -> std::string;
+public: [[nodiscard]] static auto from_string(cpp2::impl::in<std::string_view> s) -> expression_flags;
+public: [[nodiscard]] static auto from_code(cpp2::impl::in<std::string_view> s) -> expression_flags;
+
+#line 1536 "reflect.h2"
+};
+
+#line 1544 "reflect.h2"
+class regex_token
+ {
+    public: std::string string_rep; 
+
+    public: explicit regex_token(cpp2::impl::in<std::string> str);
+
+#line 1552 "reflect.h2"
+    public: explicit regex_token();
+
+#line 1557 "reflect.h2"
+    public: virtual auto generate_code([[maybe_unused]] generation_context& unnamed_param_2) const -> void = 0;
+
+    public: virtual auto add_groups([[maybe_unused]] std::set<int>& unnamed_param_2) const -> void;
+    public: [[nodiscard]] auto to_string() const& -> std::string;
+    public: auto set_string(cpp2::impl::in<std::string> s) & -> void;
+    public: virtual ~regex_token() noexcept;
+
+    public: regex_token(regex_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(regex_token const&) -> void = delete;
+
+#line 1562 "reflect.h2"
+};
+
+using token_ptr = std::shared_ptr<regex_token>;
+using token_vec = std::vector<token_ptr>;
+
+#line 1568 "reflect.h2"
+//  Adds a check in code generation.
 //
-//  regex - creates regular expressions from members
-//
-//  Each member that starts with `regex` is replaced by a regular expression
-//  of the initializer string. E.g.:
-//  ```
-//    regex := "ab";
-//  ```
-//  is replaced with
-//  ```
-//    regex := ::cpp2::regex::regular_expression<...>;
-//  ```
-//
-auto regex_gen(meta::type_declaration& t) -> void;
+class regex_token_check
+: public regex_token {
 
 #line 1574 "reflect.h2"
+    private: std::string check; 
+
+    public: explicit regex_token_check(cpp2::impl::in<std::string> str, cpp2::impl::in<std::string> check_);
+
+#line 1581 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+    public: virtual ~regex_token_check() noexcept;
+
+    public: regex_token_check(regex_token_check const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(regex_token_check const&) -> void = delete;
+
+
+#line 1584 "reflect.h2"
+};
+
+#line 1587 "reflect.h2"
+//  Adds code in code generation.
+//
+class regex_token_code
+: public regex_token {
+
+#line 1593 "reflect.h2"
+    private: std::string code; 
+
+    public: explicit regex_token_code(cpp2::impl::in<std::string> str, cpp2::impl::in<std::string> code_);
+
+#line 1600 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+    public: virtual ~regex_token_code() noexcept;
+
+    public: regex_token_code(regex_token_code const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(regex_token_code const&) -> void = delete;
+
+
+#line 1603 "reflect.h2"
+};
+
+#line 1606 "reflect.h2"
+//  Token that does not influence the matching. E.g. comment.
+//
+class regex_token_empty
+: public regex_token {
+
+#line 1612 "reflect.h2"
+    public: explicit regex_token_empty(cpp2::impl::in<std::string> str);
+
+#line 1616 "reflect.h2"
+    public: auto generate_code([[maybe_unused]] generation_context& unnamed_param_2) const -> void override;
+    public: virtual ~regex_token_empty() noexcept;
+
+    public: regex_token_empty(regex_token_empty const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(regex_token_empty const&) -> void = delete;
+
+
+#line 1619 "reflect.h2"
+};
+
+#line 1622 "reflect.h2"
+//  Represents a list of regex tokens as one token.
+//
+class regex_token_list
+: public regex_token {
+
+#line 1628 "reflect.h2"
+    public: token_vec tokens; 
+
+    public: explicit regex_token_list(cpp2::impl::in<token_vec> t);
+
+#line 1635 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+
+#line 1641 "reflect.h2"
+    public: auto add_groups(std::set<int>& groups) const -> void override;
+
+#line 1647 "reflect.h2"
+    public: [[nodiscard]] static auto gen_string(cpp2::impl::in<token_vec> vec) -> std::string;
+    public: virtual ~regex_token_list() noexcept;
+
+    public: regex_token_list(regex_token_list const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(regex_token_list const&) -> void = delete;
+
+
+#line 1654 "reflect.h2"
+};
+
+#line 1657 "reflect.h2"
+//
+//  Parse and generation context.
+//
+
+//  State of the current capturing group. See '(<pattern>)'
+//
+class parse_context_group_state
+ {
+    public: token_vec cur_match_list {};           // Current list of matchers.
+    public: token_vec alternate_match_lists {};    // List of alternate matcher lists. E.g. ab|cd|xy.
+    public: expression_flags modifiers {};         // Current modifiers for the group/regular expression.
+
+    //  Start a new alternative.
+    public: auto next_alternative() & -> void;
+
+#line 1677 "reflect.h2"
+    //  Swap this state with the other one.
+    public: auto swap(parse_context_group_state& t) & -> void;
+
+#line 1684 "reflect.h2"
+    //  Convert this state into a regex token.
+    public: [[nodiscard]] auto get_as_token() & -> token_ptr;
+
+#line 1696 "reflect.h2"
+    //  Add a token to the current matcher list.
+    public: auto add(cpp2::impl::in<token_ptr> token) & -> void;
+
+#line 1701 "reflect.h2"
+    //  True if current matcher list is empty.
+    public: [[nodiscard]] auto empty() const& -> bool;
+
+#line 1705 "reflect.h2"
+    //  Apply optimizations to the matcher list.
+    public: static auto post_process_list(token_vec& list) -> void;
+    public: parse_context_group_state(auto const& cur_match_list_, auto const& alternate_match_lists_, auto const& modifiers_);
+public: parse_context_group_state();
+
+
+#line 1719 "reflect.h2"
+};
+
+#line 1722 "reflect.h2"
+//  State for the branch reset. Takes care of the group numbering. See '(|<pattern>)'.
+//
+class parse_context_branch_reset_state
+ {
+    public: bool is_active {false}; // If we have a branch reset group.
+    public: int cur_group {1}; // Next group identifier. 0 == global capture group.
+    public: int max_group {1}; // Maximum group identifier generated.
+    public: int from {1};     // Starting identifier on new alternative branch.
+
+    //  Next group identifier.
+    public: [[nodiscard]] auto next() & -> int;
+
+#line 1740 "reflect.h2"
+    //  Set next group identifier.
+    public: auto set_next(cpp2::impl::in<int> g) & -> void;
+
+#line 1746 "reflect.h2"
+    //  Start a new alternative branch.
+    public: auto next_alternative() & -> void;
+
+#line 1753 "reflect.h2"
+    //  Initialize for a branch reset group.
+    public: auto set_active_reset(cpp2::impl::in<int> restart) & -> void;
+    public: parse_context_branch_reset_state(auto const& is_active_, auto const& cur_group_, auto const& max_group_, auto const& from_);
+public: parse_context_branch_reset_state();
+
+
+#line 1760 "reflect.h2"
+};
+
+#line 1763 "reflect.h2"
+//  Context during parsing of the regular expressions.
+// 
+//  Keeps track of the distributed group identifiers, current parsed group and branch resets.
+//
+class parse_context
+ {
+    private: std::string_view regex; // Regular expression string.
+    private: size_t pos {0};   // Current parsing position.
+    private: token_ptr root;   // Token representing the regular expression.
+
+    private: parse_context_group_state cur_group_state {}; 
+    private: parse_context_branch_reset_state cur_branch_reset_state {}; 
+
+#line 1777 "reflect.h2"
+           public: std::map<std::string,int> named_groups {}; 
+
+    private: error_func error_out; // TODO: Declaring std::function<void(std::string)> fails for cpp2.
+    private: bool has_error {false}; 
+
+    public: explicit parse_context(cpp2::impl::in<std::string_view> r, auto const& e);
+
+#line 1788 "reflect.h2"
+    //  State management functions
+    //
+
+    //  Returned group state needs to be stored and provided in `end_group`.
+    public: [[nodiscard]] auto start_group() & -> parse_context_group_state;
+
+#line 1801 "reflect.h2"
+    //  `old_state` argument needs to be from start group.
+    public: [[nodiscard]] auto end_group(cpp2::impl::in<parse_context_group_state> old_state) & -> token_ptr;
+
+#line 1809 "reflect.h2"
+    public: [[nodiscard]] auto get_modifiers() const& -> expression_flags;
+
+#line 1813 "reflect.h2"
+    public: auto set_modifiers(cpp2::impl::in<expression_flags> mod) & -> void;
+
+#line 1817 "reflect.h2"
+    //  Branch reset management functions
+    //
+
+    public: [[nodiscard]] auto branch_reset_new_state() & -> parse_context_branch_reset_state;
+
+#line 1829 "reflect.h2"
+    public: auto branch_reset_restore_state(cpp2::impl::in<parse_context_branch_reset_state> old_state) & -> void;
+
+#line 1836 "reflect.h2"
+    public: auto next_alternative() & -> void;
+
+#line 1842 "reflect.h2"
+    //  Regex token management
+    //
+    public: auto add_token(cpp2::impl::in<token_ptr> token) & -> void;
+
+#line 1848 "reflect.h2"
+    public: [[nodiscard]] auto has_token() const& -> bool;
+
+#line 1852 "reflect.h2"
+    public: [[nodiscard]] auto pop_token() & -> token_ptr;
+
+#line 1863 "reflect.h2"
+    public: [[nodiscard]] auto get_as_token() & -> token_ptr;
+
+#line 1867 "reflect.h2"
+    //  Group management
+    //
+    public: [[nodiscard]] auto get_cur_group() const& -> int;
+
+#line 1873 "reflect.h2"
+    public: [[nodiscard]] auto next_group() & -> int;
+
+#line 1877 "reflect.h2"
+    public: auto set_named_group(cpp2::impl::in<std::string> name, cpp2::impl::in<int> id) & -> void;
+
+#line 1884 "reflect.h2"
+    public: [[nodiscard]] auto get_named_group(cpp2::impl::in<std::string> name) const& -> int;
+
+#line 1895 "reflect.h2"
+    //  Position management functions
+    //
+    public: [[nodiscard]] auto current() const& -> char;
+
+    //  Get the next token in the regex, skipping spaces according to the parameters. See `x` and `xx` modifiers.
+    private: [[nodiscard]] auto get_next_position(cpp2::impl::in<bool> in_class, cpp2::impl::in<bool> no_skip) const& -> size_t;
+
+#line 1939 "reflect.h2"
+    //  Return true if next token is available.
+    private: [[nodiscard]] auto next_impl(cpp2::impl::in<bool> in_class, cpp2::impl::in<bool> no_skip) & -> bool;
+
+#line 1951 "reflect.h2"
+    public: [[nodiscard]] auto next() & -> auto;
+    public: [[nodiscard]] auto next_in_class() & -> auto;
+    public: [[nodiscard]] auto next_no_skip() & -> auto;
+
+    public: [[nodiscard]] auto next_n(cpp2::impl::in<int> n) & -> bool;
+
+#line 1964 "reflect.h2"
+    public: [[nodiscard]] auto has_next() const& -> bool;
+
+    private: [[nodiscard]] auto grab_until_impl(cpp2::impl::in<std::string> e, cpp2::impl::out<std::string> r, cpp2::impl::in<bool> any) & -> bool;
+
+#line 1987 "reflect.h2"
+    public: [[nodiscard]] auto grab_until(cpp2::impl::in<std::string> e, cpp2::impl::out<std::string> r) & -> auto;
+    public: [[nodiscard]] auto grab_until(cpp2::impl::in<char> e, cpp2::impl::out<std::string> r) & -> auto;
+    public: [[nodiscard]] auto grab_until_one_of(cpp2::impl::in<std::string> e, cpp2::impl::out<std::string> r) & -> auto;
+
+    public: [[nodiscard]] auto grab_n(cpp2::impl::in<int> n, cpp2::impl::out<std::string> r) & -> bool;
+
+#line 2004 "reflect.h2"
+    public: [[nodiscard]] auto grab_number() & -> std::string;
+
+#line 2025 "reflect.h2"
+    private: [[nodiscard]] auto peek_impl(cpp2::impl::in<bool> in_class) const& -> char;
+
+#line 2035 "reflect.h2"
+    public: [[nodiscard]] auto peek() const& -> auto;
+    public: [[nodiscard]] auto peek_in_class() const& -> auto;
+
+#line 2039 "reflect.h2"
+    //  Parsing functions
+    //
+    public: [[nodiscard]] auto parser_group_modifiers(cpp2::impl::in<std::string> change_str, expression_flags& parser_modifiers) & -> bool;
+
+#line 2095 "reflect.h2"
+    public: [[nodiscard]] auto parse_until(cpp2::impl::in<char> term) & -> bool;
+
+#line 2133 "reflect.h2"
+    public: [[nodiscard]] auto parse(cpp2::impl::in<std::string> modifiers) & -> bool;
+
+#line 2148 "reflect.h2"
+    //  Misc functions
+
+    public: [[nodiscard]] auto get_pos() const& -> auto;
+    public: [[nodiscard]] auto get_range(cpp2::impl::in<size_t> start, cpp2::impl::in<size_t> end) const& -> auto;
+    public: [[nodiscard]] auto valid() const& -> bool;
+
+    public: [[nodiscard]] auto error(cpp2::impl::in<std::string> err) & -> token_ptr;
+    public: parse_context(parse_context const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(parse_context const&) -> void = delete;
+
+
+#line 2159 "reflect.h2"
+};
+
+#line 2162 "reflect.h2"
+//  Context for one function generation. Generation of functions can be interleaved,
+//  therefore we buffer the code for one function here.
+//
+class generation_function_context {
+    public: std::string code {""}; 
+    public: std::string tabs {""}; 
+
+    public: auto add_tabs(cpp2::impl::in<int> c) & -> void;
+
+#line 2176 "reflect.h2"
+    public: auto remove_tabs(cpp2::impl::in<int> c) & -> void;
+    public: generation_function_context(auto const& code_, auto const& tabs_);
+public: generation_function_context();
+
+
+#line 2179 "reflect.h2"
+};
+
+#line 2182 "reflect.h2"
+//  Context for generating the state machine.
+class generation_context
+ {
+    private: std::vector<generation_function_context> gen_stack {1}; // Element 0 contains all the code.
+
+    private: int matcher_func {0}; 
+    private: int reset_func {0}; 
+    private: int temp_name {0}; 
+    private: std::string entry_func {""}; 
+
+    //  Generation helpers
+    //
+    public: [[nodiscard]] auto match_parameters() const& -> std::string;
+
+    //  Code generation.
+
+    //  Add code line.
+    public: auto add(cpp2::impl::in<std::string> s) & -> void;
+
+#line 2204 "reflect.h2"
+    //  Add check for token. The check needs to be a function call that returns a boolean.
+    public: auto add_check(cpp2::impl::in<std::string> check) & -> void;
+
+#line 2210 "reflect.h2"
+    //  Add a stateful check. The check needs to return a `match_return`.
+    public: auto add_statefull(cpp2::impl::in<std::string> next_func, cpp2::impl::in<std::string> check) & -> void;
+
+#line 2219 "reflect.h2"
+    protected: auto start_func_named(cpp2::impl::in<std::string> name) & -> void;
+
+#line 2230 "reflect.h2"
+    protected: [[nodiscard]] auto start_func() & -> std::string;
+
+#line 2237 "reflect.h2"
+    protected: auto end_func_statefull(cpp2::impl::in<std::string> s) & -> void;
+
+#line 2256 "reflect.h2"
+    //  Generate the function for a token.
+    public: [[nodiscard]] auto generate_func(cpp2::impl::in<token_ptr> token) & -> std::string;
+
+#line 2266 "reflect.h2"
+    //  Generate the reset for a list of group identifiers.
+    public: [[nodiscard]] auto generate_reset(cpp2::impl::in<std::set<int>> groups) & -> std::string;
+
+#line 2289 "reflect.h2"
+    //  Name generation
+    //
+    protected: [[nodiscard]] auto gen_func_name() & -> std::string;
+
+#line 2297 "reflect.h2"
+    public: [[nodiscard]] auto next_func_name() & -> std::string;
+
+#line 2301 "reflect.h2"
+    protected: [[nodiscard]] auto gen_reset_func_name() & -> std::string;
+
+#line 2307 "reflect.h2"
+    public: [[nodiscard]] auto gen_temp() & -> std::string;
+
+#line 2313 "reflect.h2"
+    //  Context management
+    //
+    public: [[nodiscard]] auto new_context() & -> generation_function_context*;
+
+#line 2323 "reflect.h2"
+    public: auto finish_context() & -> void;
+
+#line 2331 "reflect.h2"
+    //  Misc functions
+    //
+    private: [[nodiscard]] auto get_current() & -> generation_function_context*;
+
+#line 2337 "reflect.h2"
+    private: [[nodiscard]] auto get_base() & -> generation_function_context*;
+
+#line 2341 "reflect.h2"
+    public: [[nodiscard]] auto get_entry_func() const& -> std::string;
+
+#line 2345 "reflect.h2"
+    public: [[nodiscard]] auto create_named_group_lookup(cpp2::impl::in<std::map<std::string,int>> named_groups) const& -> std::string;
+
+#line 2369 "reflect.h2"
+    //  Run the generation for the token.
+    public: [[nodiscard]] auto run(cpp2::impl::in<token_ptr> token) & -> std::string;
+    public: generation_context() = default;
+    public: generation_context(generation_context const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(generation_context const&) -> void = delete;
+
+
+#line 2375 "reflect.h2"
+};
+
+//  Regex syntax: |  Example: ab|ba
+//
+//  Non greedy implementation. First alternative that matches is chosen.
+//
+class alternative_token
+: public regex_token_empty {
+                                     // No code gen here. alternative_token_gen is created in the parse_context
+
+    public: explicit alternative_token();
+
+    public: [[nodiscard]] static auto parse(parse_context& ctx) -> token_ptr;
+    public: virtual ~alternative_token() noexcept;
+
+    public: alternative_token(alternative_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(alternative_token const&) -> void = delete;
+
+
+#line 2394 "reflect.h2"
+};
+
+class alternative_token_gen
+: public regex_token {
+
+#line 2400 "reflect.h2"
+    private: token_vec alternatives; 
+
+    public: explicit alternative_token_gen(cpp2::impl::in<token_vec> a);
+
+#line 2407 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+
+#line 2424 "reflect.h2"
+    public: auto add_groups(std::set<int>& groups) const -> void override;
+
+#line 2431 "reflect.h2"
+    public: [[nodiscard]] static auto gen_string(cpp2::impl::in<token_vec> a) -> std::string;
+    public: virtual ~alternative_token_gen() noexcept;
+
+    public: alternative_token_gen(alternative_token_gen const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(alternative_token_gen const&) -> void = delete;
+
+
+#line 2443 "reflect.h2"
+};
+
+#line 2446 "reflect.h2"
+//  Regex syntax: .
+//
+class any_token
+: public regex_token_check {
+
+#line 2452 "reflect.h2"
+    public: explicit any_token(cpp2::impl::in<bool> single_line);
+
+#line 2456 "reflect.h2"
+    public: [[nodiscard]] static auto parse(parse_context& ctx) -> token_ptr;
+    public: virtual ~any_token() noexcept;
+
+    public: any_token(any_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(any_token const&) -> void = delete;
+
+
+#line 2461 "reflect.h2"
+};
+
+#line 2464 "reflect.h2"
+//  Regex syntax: a
+//
+class char_token
+: public regex_token {
+
+#line 2470 "reflect.h2"
+    private: std::string token; 
+    private: bool ignore_case; 
+
+    public: explicit char_token(cpp2::impl::in<char> t, cpp2::impl::in<bool> ignore_case_);
+
+#line 2479 "reflect.h2"
+    public: [[nodiscard]] static auto parse(parse_context& ctx) -> token_ptr;
+
+#line 2483 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+
+#line 2506 "reflect.h2"
+    public: auto gen_case_insensitive(cpp2::impl::in<std::string> lower, cpp2::impl::in<std::string> upper, generation_context& ctx) const& -> void;
+
+#line 2527 "reflect.h2"
+    public: auto gen_case_sensitive(generation_context& ctx) const& -> void;
+
+#line 2545 "reflect.h2"
+    public: [[nodiscard]] auto add_escapes(std::string str) const& -> std::string;
+
+#line 2560 "reflect.h2"
+    public: auto append(char_token const& that) & -> void;
+    public: virtual ~char_token() noexcept;
+
+    public: char_token(char_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(char_token const&) -> void = delete;
+
+
+#line 2564 "reflect.h2"
+};
+
+#line 2567 "reflect.h2"
+//  Regex syntax: [<character classes>]  Example: [abcx-y[:digits:]]
+//
+class class_token
+: public regex_token {
+
+#line 2573 "reflect.h2"
+    private: bool negate; 
+    private: bool case_insensitive; 
+    private: std::string class_str; 
+
+    public: explicit class_token(cpp2::impl::in<bool> negate_, cpp2::impl::in<bool> case_insensitive_, cpp2::impl::in<std::string> class_str_, cpp2::impl::in<std::string> str);
+
+#line 2585 "reflect.h2"
+    //  TODO: Rework class generation: Generate check functions for classes.
+    public: [[nodiscard]] static auto parse(parse_context& ctx) -> token_ptr;
+
+#line 2711 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+
+#line 2716 "reflect.h2"
+    private: [[nodiscard]] static auto create_matcher(cpp2::impl::in<std::string> name, cpp2::impl::in<std::string> template_arguments) -> std::string;
+    public: virtual ~class_token() noexcept;
+
+    public: class_token(class_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(class_token const&) -> void = delete;
+
+
+#line 2723 "reflect.h2"
+};
+
+#line 2726 "reflect.h2"
+//  Regex syntax: \a or \n or \[
+//
+[[nodiscard]] auto escape_token_parse(parse_context& ctx) -> token_ptr;
+
+#line 2767 "reflect.h2"
+//  Regex syntax: \K Example: ab\Kcd
+//
+[[nodiscard]] auto global_group_reset_token_parse(parse_context& ctx) -> token_ptr;
+
+#line 2778 "reflect.h2"
+//  Regex syntax: \<number>  Example: \1
+//                \g{name_or_number}
+//                \k{name_or_number}
+//                \k<name_or_number>
+//                \k'name_or_number'
+//
+class group_ref_token
+: public regex_token {
+
+#line 2788 "reflect.h2"
+    private: int id; 
+    private: bool case_insensitive; 
+
+    public: explicit group_ref_token(cpp2::impl::in<int> id_, cpp2::impl::in<bool> case_insensitive_, cpp2::impl::in<std::string> str);
+
+#line 2798 "reflect.h2"
+    public: [[nodiscard]] static auto parse(parse_context& ctx) -> token_ptr;
+
+#line 2899 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+    public: virtual ~group_ref_token() noexcept;
+
+    public: group_ref_token(group_ref_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(group_ref_token const&) -> void = delete;
+
+
+#line 2902 "reflect.h2"
+};
+
+#line 2905 "reflect.h2"
+//  Regex syntax: (<tokens>)      Example: (abc)
+//                (?<modifiers)            (?i)
+//                (?<modifiers>:<tokens>)  (?i:abc)
+//                (?<<name>>:<tokens>)     (?<start>:abc)
+//                (?#<comment>)            (#Step 1 finished)
+//                (?|<tokens>)             (?|(abc)|(cde))
+//                (?=<tokens>)             (?=abc)
+//                (?!<tokens>)             (?!abc)
+//                (*<name>:<tokens)        (*pla:abc)
+//
+class group_token
+: public regex_token {
+
+#line 2919 "reflect.h2"
+    private: int number {-1}; 
+    private: token_ptr inner {nullptr}; 
+
+    public: [[nodiscard]] static auto parse_lookahead(parse_context& ctx, cpp2::impl::in<std::string> syntax, cpp2::impl::in<bool> positive) -> token_ptr;
+
+#line 2936 "reflect.h2"
+    public: [[nodiscard]] static auto parse(parse_context& ctx) -> token_ptr;
+
+#line 3073 "reflect.h2"
+    public: [[nodiscard]] static auto gen_string(cpp2::impl::in<std::string> name, cpp2::impl::in<bool> name_brackets, cpp2::impl::in<bool> has_modifier, cpp2::impl::in<std::string> modifiers, cpp2::impl::in<token_ptr> inner_) -> std::string;
+
+#line 3091 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+
+#line 3111 "reflect.h2"
+    public: auto add_groups(std::set<int>& groups) const -> void override;
+    public: virtual ~group_token() noexcept;
+
+    public: group_token() = default;
+    public: group_token(group_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(group_token const&) -> void = delete;
+
+
+#line 3118 "reflect.h2"
+};
+
+#line 3121 "reflect.h2"
+//  Regex syntax: \x<number> or \x{<number>}  Example: \x{62}
+//
+[[nodiscard]] auto hexadecimal_token_parse(parse_context& ctx) -> token_ptr;
+
+#line 3162 "reflect.h2"
+//  Regex syntax: $  Example: aa$
+//
+[[nodiscard]] auto line_end_token_parse(parse_context& ctx) -> token_ptr;
+
+#line 3182 "reflect.h2"
+//  Regex syntax: ^  Example: ^aa
+//
+[[nodiscard]] auto line_start_token_parse(parse_context& ctx) -> token_ptr;
+
+#line 3198 "reflect.h2"
+//  Regex syntax: (?=) or (?!) or (*pla), etc.  Example: (?=AA)
+//
+//  Parsed in group_token.
+//
+class lookahead_token
+: public regex_token {
+
+#line 3206 "reflect.h2"
+    protected: bool positive; 
+    public:   token_ptr inner {nullptr}; 
+
+    public: explicit lookahead_token(cpp2::impl::in<bool> positive_);
+
+#line 3213 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+
+#line 3219 "reflect.h2"
+    public: auto add_groups(std::set<int>& groups) const -> void override;
+    public: virtual ~lookahead_token() noexcept;
+
+    public: lookahead_token(lookahead_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(lookahead_token const&) -> void = delete;
+
+
+#line 3222 "reflect.h2"
+};
+
+#line 3225 "reflect.h2"
+// Named character classes
+//
+[[nodiscard]] auto named_class_token_parse(parse_context& ctx) -> token_ptr;
+
+#line 3253 "reflect.h2"
+//  Regex syntax: \o{<number>}  Example: \o{142}
+//
+[[nodiscard]] auto octal_token_parse(parse_context& ctx) -> token_ptr;
+
+#line 3281 "reflect.h2"
+//  Regex syntax: <matcher>{min, max}  Example: a{2,4}
+//
+class range_token
+: public regex_token {
+
+#line 3287 "reflect.h2"
+              protected: int min_count {-1}; 
+              protected: int max_count {-1}; 
+              protected: int kind {range_flags::greedy}; 
+              protected: token_ptr inner_token {nullptr}; 
+
+    public: explicit range_token();
+
+    public: [[nodiscard]] static auto parse(parse_context& ctx) -> token_ptr;
+
+#line 3367 "reflect.h2"
+    public: auto parse_modifier(parse_context& ctx) & -> void;
+
+#line 3379 "reflect.h2"
+    public: [[nodiscard]] auto gen_mod_string() const& -> std::string;
+
+#line 3392 "reflect.h2"
+    public: [[nodiscard]] auto gen_range_string() const& -> std::string;
+
+#line 3411 "reflect.h2"
+    public: auto generate_code(generation_context& ctx) const -> void override;
+
+#line 3422 "reflect.h2"
+    public: auto add_groups(std::set<int>& groups) const -> void override;
+    public: virtual ~range_token() noexcept;
+
+    public: range_token(range_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(range_token const&) -> void = delete;
+
+
+#line 3426 "reflect.h2"
+};
+
+#line 3429 "reflect.h2"
+//  Regex syntax: *, +, or ?  Example: aa*
+//
+class special_range_token
+: public range_token {
+
+#line 3435 "reflect.h2"
+    public: [[nodiscard]] static auto parse(parse_context& ctx) -> token_ptr;
+    public: virtual ~special_range_token() noexcept;
+
+    public: special_range_token() = default;
+    public: special_range_token(special_range_token const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(special_range_token const&) -> void = delete;
+
+
+#line 3465 "reflect.h2"
+};
+
+#line 3468 "reflect.h2"
+//  Regex syntax: \b or \B  Example: \bword\b
+//
+//  Matches the start end end of word boundaries.
+//
+[[nodiscard]] auto word_boundary_token_parse(parse_context& ctx) -> token_ptr;
+
+#line 3490 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//  Parser for regular expression.
+//
+//-----------------------------------------------------------------------
+//
+
+//  Parser and generator for regular expressions.
+template<typename Error_out> class regex_generator
+ {
+    private: std::string_view regex; 
+    private: std::string modifier {""}; 
+    private: std::string modifier_escape {""}; 
+
+    private: Error_out error_out; 
+
+    private: std::string source {""}; 
+
+    public: explicit regex_generator(cpp2::impl::in<std::string_view> r, Error_out const& e);
+
+#line 3513 "reflect.h2"
+    public: [[nodiscard]] auto parse() & -> std::string;
+
+#line 3548 "reflect.h2"
+    private: auto extract_modifiers() & -> void;
+    public: regex_generator(regex_generator const&) = delete; /* No 'that' constructor, suppress copy */
+    public: auto operator=(regex_generator const&) -> void = delete;
+
+
+#line 3562 "reflect.h2"
+};
+
+template<typename Err> [[nodiscard]] auto generate_regex(cpp2::impl::in<std::string_view> regex, Err const& err) -> std::string;
+
+#line 3572 "reflect.h2"
+} // end regex subnamespace
+
+#line 3575 "reflect.h2"
+namespace meta {
+
+auto regex_gen(meta::type_declaration& t) -> void;
+
+#line 3630 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //  apply_metafunctions
@@ -796,7 +1431,7 @@ auto regex_gen(meta::type_declaration& t) -> void;
     auto const& error
     ) -> bool;
 
-#line 1693 "reflect.h2"
+#line 3749 "reflect.h2"
 }
 
 }
@@ -806,12 +1441,27 @@ auto regex_gen(meta::type_declaration& t) -> void;
 
 #line 1 "reflect.h2"
 
-#line 22 "reflect.h2"
+#line 21 "reflect.h2"
 namespace cpp2 {
 
 namespace meta {
 
-#line 48 "reflect.h2"
+#line 26 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//  Compiler services
+//
+//-----------------------------------------------------------------------
+//
+
+#line 35 "reflect.h2"
+    //  Common data members
+    //
+
+#line 45 "reflect.h2"
+    //  Constructor
+    //
+#line 47 "reflect.h2"
     compiler_services::compiler_services(
 
         std::vector<error_entry>* errors_, 
@@ -821,23 +1471,25 @@ namespace meta {
         , errors_original_size{ cpp2::unsafe_narrow<int>(std::ssize(*cpp2::impl::assert_not_null(errors))) }
         , generated_tokens{ generated_tokens_ }
         , parser{ *cpp2::impl::assert_not_null(errors) }
-#line 53 "reflect.h2"
+#line 52 "reflect.h2"
     {
 
-#line 58 "reflect.h2"
+#line 57 "reflect.h2"
     }
 
-#line 62 "reflect.h2"
+    //  Common API
+    //
+#line 61 "reflect.h2"
     auto compiler_services::set_metafunction_name(cpp2::impl::in<std::string_view> name, cpp2::impl::in<std::vector<std::string>> args) & -> void{
         metafunction_name  = name;
         metafunction_args  = args;
         metafunctions_used = CPP2_UFCS(empty)(args);
     }
 
-#line 68 "reflect.h2"
+#line 67 "reflect.h2"
     [[nodiscard]] auto compiler_services::get_metafunction_name() const& -> std::string_view { return metafunction_name;  }
 
-#line 70 "reflect.h2"
+#line 69 "reflect.h2"
     [[nodiscard]] auto compiler_services::get_argument(cpp2::impl::in<int> index) & -> std::string{
         metafunctions_used = true;
         if (([_0 = 0, _1 = index, _2 = CPP2_UFCS(ssize)(metafunction_args)]{ return cpp2::impl::cmp_less_eq(_0,_1) && cpp2::impl::cmp_less(_1,_2); }())) {
@@ -846,16 +1498,16 @@ namespace meta {
         return ""; 
     }
 
-#line 78 "reflect.h2"
+#line 77 "reflect.h2"
     [[nodiscard]] auto compiler_services::get_arguments() & -> std::vector<std::string>{
         metafunctions_used = true;
         return metafunction_args; 
     }
 
-#line 83 "reflect.h2"
+#line 82 "reflect.h2"
     [[nodiscard]] auto compiler_services::arguments_were_used() const& -> bool { return metafunctions_used;  }
 
-#line 85 "reflect.h2"
+#line 84 "reflect.h2"
     [[nodiscard]] auto compiler_services::parse_statement(
 
         std::string_view source
@@ -863,7 +1515,7 @@ namespace meta {
 
     {
             cpp2::impl::deferred_init<std::unique_ptr<statement_node>> ret;
-#line 91 "reflect.h2"
+#line 90 "reflect.h2"
         auto original_source {source}; 
 
         CPP2_UFCS(push_back)(generated_lines, std::vector<source_line>());
@@ -878,7 +1530,7 @@ auto newline_pos{CPP2_UFCS(find)(source, '\n')};
         //  First split this string into source_lines
         //
 
-#line 103 "reflect.h2"
+#line 102 "reflect.h2"
         if ( cpp2::impl::cmp_greater(CPP2_UFCS(ssize)(source),1) 
             && newline_pos != source.npos) 
         {
@@ -891,7 +1543,7 @@ auto newline_pos{CPP2_UFCS(find)(source, '\n')};
         }
 }
 
-#line 114 "reflect.h2"
+#line 113 "reflect.h2"
         if (!(CPP2_UFCS(empty)(source))) {
             cpp2::move(add_line)(cpp2::move(source));
         }
@@ -916,14 +1568,17 @@ auto newline_pos{CPP2_UFCS(find)(source, '\n')};
         }return std::move(ret.value()); 
     }
 
-#line 138 "reflect.h2"
+#line 137 "reflect.h2"
     [[nodiscard]] auto compiler_services::position() const -> source_position
 
     {
         return {  }; 
     }
 
-#line 147 "reflect.h2"
+    //  Error diagnosis and handling, integrated with compiler output
+    //  Unlike a contract violation, .requires continues further processing
+    //
+#line 146 "reflect.h2"
     auto compiler_services::require(
 
         cpp2::impl::in<bool> b, 
@@ -935,7 +1590,7 @@ auto newline_pos{CPP2_UFCS(find)(source, '\n')};
         }
     }
 
-#line 158 "reflect.h2"
+#line 157 "reflect.h2"
     auto compiler_services::error(cpp2::impl::in<std::string_view> msg) const& -> void
     {
         auto message {cpp2::impl::as_<std::string>(msg)}; 
@@ -945,7 +1600,10 @@ auto newline_pos{CPP2_UFCS(find)(source, '\n')};
         static_cast<void>(CPP2_UFCS(emplace_back)((*cpp2::impl::assert_not_null(errors)), position(), cpp2::move(message)));
     }
 
-#line 170 "reflect.h2"
+    //  Enable custom contracts on this object, integrated with compiler output
+    //  Unlike .requires, a contract violation stops further processing
+    //
+#line 169 "reflect.h2"
     auto compiler_services::report_violation(auto const& msg) const& -> void{
         error(msg);
         throw(std::runtime_error(
@@ -954,7 +1612,7 @@ auto newline_pos{CPP2_UFCS(find)(source, '\n')};
         ));
     }
 
-#line 178 "reflect.h2"
+#line 177 "reflect.h2"
     [[nodiscard]] auto compiler_services::is_active() const& -> auto { return true;  }
 
     compiler_services::~compiler_services() noexcept{}
@@ -967,7 +1625,55 @@ compiler_services::compiler_services(compiler_services const& that)
                                 , metafunction_args{ that.metafunction_args }
                                 , metafunctions_used{ that.metafunctions_used }{}
 
-#line 234 "reflect.h2"
+#line 181 "reflect.h2"
+/*
+//-----------------------------------------------------------------------
+//
+//  Type IDs
+//
+//-----------------------------------------------------------------------
+//
+
+//  All type_ids are wrappers around a pointer to node
+//
+type_id: @polymorphic_base @copyable type =
+{
+    this: compiler_services = ();
+
+    n: type_id_node;
+
+    protected operator=: (
+        out this,
+        n_: type_id_node,
+        s : compiler_services
+    )
+    = {
+        compiler_services = s;
+        n = n_;
+        assert( n, "a meta::type_id must point to a valid type_id_node, not null" );
+    }
+
+    is_wildcard         : (this) -> bool        = n.is_wildcard();
+    is_pointer_qualified: (this) -> bool        = n.is_pointer_qualified();
+    template_args_count : (this) -> int         = n.template_arguments().ssize();
+    to_string           : (this) -> std::string = n.to_string();
+
+    position: (override this) -> source_position = n.position();
+}
+*/
+
+#line 218 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//  Declarations
+//
+//-----------------------------------------------------------------------
+//
+
+//  All declarations are wrappers around a pointer to node
+//
+
+#line 233 "reflect.h2"
     declaration_base::declaration_base(
 
         declaration_node* n_, 
@@ -975,17 +1681,17 @@ compiler_services::compiler_services(compiler_services const& that)
     )
         : compiler_services{ s }
         , n{ n_ }
-#line 239 "reflect.h2"
+#line 238 "reflect.h2"
     {
 
-#line 242 "reflect.h2"
+#line 241 "reflect.h2"
         if (cpp2::cpp2_default.is_active() && !(n) ) { cpp2::cpp2_default.report_violation(CPP2_CONTRACT_MSG("a meta::declaration must point to a valid declaration_node, not null")); }
     }
 
-#line 245 "reflect.h2"
+#line 244 "reflect.h2"
     [[nodiscard]] auto declaration_base::position() const -> source_position { return CPP2_UFCS(position)((*cpp2::impl::assert_not_null(n)));  }
 
-#line 247 "reflect.h2"
+#line 246 "reflect.h2"
     [[nodiscard]] auto declaration_base::print() const& -> std::string { return CPP2_UFCS(pretty_print_visualize)((*cpp2::impl::assert_not_null(n)), 0);  }
 
     declaration_base::~declaration_base() noexcept{}
@@ -993,121 +1699,126 @@ declaration_base::declaration_base(declaration_base const& that)
                                 : compiler_services{ static_cast<compiler_services const&>(that) }
                                 , n{ that.n }{}
 
-#line 258 "reflect.h2"
+#line 250 "reflect.h2"
+//-----------------------------------------------------------------------
+//  All declarations
+//
+
+#line 257 "reflect.h2"
     declaration::declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     )
         : declaration_base{ n_, s }
-#line 263 "reflect.h2"
+#line 262 "reflect.h2"
     {
 
     }
 
-#line 267 "reflect.h2"
+#line 266 "reflect.h2"
     [[nodiscard]] auto declaration::is_public() const& -> bool { return CPP2_UFCS(is_public)((*cpp2::impl::assert_not_null(n))); }
-#line 268 "reflect.h2"
+#line 267 "reflect.h2"
     [[nodiscard]] auto declaration::is_protected() const& -> bool { return CPP2_UFCS(is_protected)((*cpp2::impl::assert_not_null(n))); }
-#line 269 "reflect.h2"
+#line 268 "reflect.h2"
     [[nodiscard]] auto declaration::is_private() const& -> bool { return CPP2_UFCS(is_private)((*cpp2::impl::assert_not_null(n))); }
-#line 270 "reflect.h2"
+#line 269 "reflect.h2"
     [[nodiscard]] auto declaration::is_default_access() const& -> bool { return CPP2_UFCS(is_default_access)((*cpp2::impl::assert_not_null(n)));  }
 
-#line 272 "reflect.h2"
+#line 271 "reflect.h2"
     auto declaration::default_to_public() & -> void { static_cast<void>(CPP2_UFCS(make_public)((*cpp2::impl::assert_not_null(n)))); }
-#line 273 "reflect.h2"
+#line 272 "reflect.h2"
     auto declaration::default_to_protected() & -> void { static_cast<void>(CPP2_UFCS(make_protected)((*cpp2::impl::assert_not_null(n))));  }
-#line 274 "reflect.h2"
+#line 273 "reflect.h2"
     auto declaration::default_to_private() & -> void { static_cast<void>(CPP2_UFCS(make_private)((*cpp2::impl::assert_not_null(n)))); }
 
-#line 276 "reflect.h2"
+#line 275 "reflect.h2"
     [[nodiscard]] auto declaration::make_public() & -> bool { return CPP2_UFCS(make_public)((*cpp2::impl::assert_not_null(n))); }
-#line 277 "reflect.h2"
+#line 276 "reflect.h2"
     [[nodiscard]] auto declaration::make_protected() & -> bool { return CPP2_UFCS(make_protected)((*cpp2::impl::assert_not_null(n))); }
-#line 278 "reflect.h2"
+#line 277 "reflect.h2"
     [[nodiscard]] auto declaration::make_private() & -> bool { return CPP2_UFCS(make_private)((*cpp2::impl::assert_not_null(n))); }
 
-#line 280 "reflect.h2"
+#line 279 "reflect.h2"
     [[nodiscard]] auto declaration::has_name() const& -> bool { return CPP2_UFCS(has_name)((*cpp2::impl::assert_not_null(n))); }
-#line 281 "reflect.h2"
+#line 280 "reflect.h2"
     [[nodiscard]] auto declaration::has_name(cpp2::impl::in<std::string_view> s) const& -> bool { return CPP2_UFCS(has_name)((*cpp2::impl::assert_not_null(n)), s); }
 
-#line 283 "reflect.h2"
+#line 282 "reflect.h2"
     [[nodiscard]] auto declaration::name() const& -> std::string_view{
         if (has_name()) {return CPP2_UFCS(as_string_view)((*cpp2::impl::assert_not_null(CPP2_UFCS(name)(*cpp2::impl::assert_not_null(n))))); }
         else          { return ""; }
     }
 
-#line 288 "reflect.h2"
+#line 287 "reflect.h2"
     [[nodiscard]] auto declaration::has_initializer() const& -> bool { return CPP2_UFCS(has_initializer)((*cpp2::impl::assert_not_null(n)));  }
 
-#line 290 "reflect.h2"
+#line 289 "reflect.h2"
     [[nodiscard]] auto declaration::is_global() const& -> bool { return CPP2_UFCS(is_global)((*cpp2::impl::assert_not_null(n))); }
-#line 291 "reflect.h2"
+#line 290 "reflect.h2"
     [[nodiscard]] auto declaration::is_function() const& -> bool { return CPP2_UFCS(is_function)((*cpp2::impl::assert_not_null(n))); }
-#line 292 "reflect.h2"
+#line 291 "reflect.h2"
     [[nodiscard]] auto declaration::is_object() const& -> bool { return CPP2_UFCS(is_object)((*cpp2::impl::assert_not_null(n))); }
-#line 293 "reflect.h2"
+#line 292 "reflect.h2"
     [[nodiscard]] auto declaration::is_base_object() const& -> bool { return CPP2_UFCS(is_base_object)((*cpp2::impl::assert_not_null(n))); }
-#line 294 "reflect.h2"
+#line 293 "reflect.h2"
     [[nodiscard]] auto declaration::is_member_object() const& -> bool { return CPP2_UFCS(is_member_object)((*cpp2::impl::assert_not_null(n)));  }
-#line 295 "reflect.h2"
+#line 294 "reflect.h2"
     [[nodiscard]] auto declaration::is_type() const& -> bool { return CPP2_UFCS(is_type)((*cpp2::impl::assert_not_null(n))); }
-#line 296 "reflect.h2"
+#line 295 "reflect.h2"
     [[nodiscard]] auto declaration::is_namespace() const& -> bool { return CPP2_UFCS(is_namespace)((*cpp2::impl::assert_not_null(n))); }
-#line 297 "reflect.h2"
+#line 296 "reflect.h2"
     [[nodiscard]] auto declaration::is_alias() const& -> bool { return CPP2_UFCS(is_alias)((*cpp2::impl::assert_not_null(n))); }
 
-#line 299 "reflect.h2"
+#line 298 "reflect.h2"
     [[nodiscard]] auto declaration::is_type_alias() const& -> bool { return CPP2_UFCS(is_type_alias)((*cpp2::impl::assert_not_null(n))); }
-#line 300 "reflect.h2"
+#line 299 "reflect.h2"
     [[nodiscard]] auto declaration::is_namespace_alias() const& -> bool { return CPP2_UFCS(is_namespace_alias)((*cpp2::impl::assert_not_null(n)));  }
-#line 301 "reflect.h2"
+#line 300 "reflect.h2"
     [[nodiscard]] auto declaration::is_object_alias() const& -> bool { return CPP2_UFCS(is_object_alias)((*cpp2::impl::assert_not_null(n))); }
 
-#line 303 "reflect.h2"
+#line 302 "reflect.h2"
     [[nodiscard]] auto declaration::is_function_expression() const& -> bool { return CPP2_UFCS(is_function_expression)((*cpp2::impl::assert_not_null(n)));  }
 
-#line 305 "reflect.h2"
+#line 304 "reflect.h2"
     [[nodiscard]] auto declaration::as_function() const& -> function_declaration { return function_declaration(n, (*this));  }
-#line 306 "reflect.h2"
+#line 305 "reflect.h2"
     [[nodiscard]] auto declaration::as_object() const& -> object_declaration { return object_declaration(n, (*this)); }
-#line 307 "reflect.h2"
+#line 306 "reflect.h2"
     [[nodiscard]] auto declaration::as_type() const& -> type_declaration { return type_declaration(n, (*this)); }
-#line 308 "reflect.h2"
+#line 307 "reflect.h2"
     [[nodiscard]] auto declaration::as_alias() const& -> alias_declaration { return alias_declaration(n, (*this)); }
 
-#line 310 "reflect.h2"
+#line 309 "reflect.h2"
     [[nodiscard]] auto declaration::get_parent() const& -> declaration { return declaration((*cpp2::impl::assert_not_null(n)).parent_declaration, (*this)); }
 
-#line 312 "reflect.h2"
+#line 311 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_function() const& -> bool { return CPP2_UFCS(parent_is_function)((*cpp2::impl::assert_not_null(n))); }
-#line 313 "reflect.h2"
+#line 312 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_object() const& -> bool { return CPP2_UFCS(parent_is_object)((*cpp2::impl::assert_not_null(n))); }
-#line 314 "reflect.h2"
+#line 313 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_type() const& -> bool { return CPP2_UFCS(parent_is_type)((*cpp2::impl::assert_not_null(n))); }
-#line 315 "reflect.h2"
+#line 314 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_namespace() const& -> bool { return CPP2_UFCS(parent_is_namespace)((*cpp2::impl::assert_not_null(n))); }
-#line 316 "reflect.h2"
+#line 315 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_alias() const& -> bool { return CPP2_UFCS(parent_is_alias)((*cpp2::impl::assert_not_null(n))); }
 
-#line 318 "reflect.h2"
+#line 317 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_type_alias() const& -> bool { return CPP2_UFCS(parent_is_type_alias)((*cpp2::impl::assert_not_null(n))); }
-#line 319 "reflect.h2"
+#line 318 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_namespace_alias() const& -> bool { return CPP2_UFCS(parent_is_namespace_alias)((*cpp2::impl::assert_not_null(n)));  }
-#line 320 "reflect.h2"
+#line 319 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_object_alias() const& -> bool { return CPP2_UFCS(parent_is_object_alias)((*cpp2::impl::assert_not_null(n))); }
 
-#line 322 "reflect.h2"
+#line 321 "reflect.h2"
     [[nodiscard]] auto declaration::parent_is_polymorphic() const& -> bool { return CPP2_UFCS(parent_is_polymorphic)((*cpp2::impl::assert_not_null(n)));  }
 
-#line 324 "reflect.h2"
+#line 323 "reflect.h2"
     auto declaration::mark_for_removal_from_enclosing_type() & -> void
-
+                                                    // this precondition should be sufficient ...
     {
         if (cpp2::type_safety.is_active() && !(parent_is_type()) ) { cpp2::type_safety.report_violation(""); }
-#line 327 "reflect.h2"
+#line 326 "reflect.h2"
         auto test {CPP2_UFCS(type_member_mark_for_removal)((*cpp2::impl::assert_not_null(n)))}; 
         if (cpp2::cpp2_default.is_active() && !(cpp2::move(test)) ) { cpp2::cpp2_default.report_violation(""); }// ... to ensure this assert is true
     }
@@ -1116,88 +1827,93 @@ declaration_base::declaration_base(declaration_base const& that)
 declaration::declaration(declaration const& that)
                                 : declaration_base{ static_cast<declaration_base const&>(that) }{}
 
-#line 340 "reflect.h2"
+#line 332 "reflect.h2"
+//-----------------------------------------------------------------------
+//  Function declarations
+//
+
+#line 339 "reflect.h2"
     function_declaration::function_declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     )
         : declaration{ n_, s }
-#line 345 "reflect.h2"
+#line 344 "reflect.h2"
     {
 
         if (cpp2::cpp2_default.is_active() && !(CPP2_UFCS(is_function)((*cpp2::impl::assert_not_null(n)))) ) { cpp2::cpp2_default.report_violation(""); }
     }
 
-#line 350 "reflect.h2"
+#line 349 "reflect.h2"
     [[nodiscard]] auto function_declaration::index_of_parameter_named(cpp2::impl::in<std::string_view> s) const& -> int { return CPP2_UFCS(index_of_parameter_named)((*cpp2::impl::assert_not_null(n)), s); }
-#line 351 "reflect.h2"
+#line 350 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool { return CPP2_UFCS(has_parameter_named)((*cpp2::impl::assert_not_null(n)), s); }
-#line 352 "reflect.h2"
+#line 351 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_in_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool { return CPP2_UFCS(has_in_parameter_named)((*cpp2::impl::assert_not_null(n)), s); }
-#line 353 "reflect.h2"
+#line 352 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_copy_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool { return CPP2_UFCS(has_copy_parameter_named)((*cpp2::impl::assert_not_null(n)), s); }
-#line 354 "reflect.h2"
+#line 353 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_inout_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool { return CPP2_UFCS(has_inout_parameter_named)((*cpp2::impl::assert_not_null(n)), s); }
-#line 355 "reflect.h2"
+#line 354 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_out_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool { return CPP2_UFCS(has_out_parameter_named)((*cpp2::impl::assert_not_null(n)), s); }
-#line 356 "reflect.h2"
+#line 355 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_move_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool { return CPP2_UFCS(has_move_parameter_named)((*cpp2::impl::assert_not_null(n)), s); }
-#line 357 "reflect.h2"
+#line 356 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_forward_parameter_named(cpp2::impl::in<std::string_view> s) const& -> bool { return CPP2_UFCS(has_forward_parameter_named)((*cpp2::impl::assert_not_null(n)), s); }
-#line 358 "reflect.h2"
+#line 357 "reflect.h2"
     [[nodiscard]] auto function_declaration::first_parameter_name() const& -> std::string { return CPP2_UFCS(first_parameter_name)((*cpp2::impl::assert_not_null(n))); }
 
-#line 360 "reflect.h2"
+#line 359 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_parameter_with_name_and_pass(cpp2::impl::in<std::string_view> s, cpp2::impl::in<passing_style> pass) const& -> bool { 
                                                   return CPP2_UFCS(has_parameter_with_name_and_pass)((*cpp2::impl::assert_not_null(n)), s, pass);  }
-#line 362 "reflect.h2"
+#line 361 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_function_with_this() const& -> bool { return CPP2_UFCS(is_function_with_this)((*cpp2::impl::assert_not_null(n))); }
-#line 363 "reflect.h2"
+#line 362 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_virtual() const& -> bool { return CPP2_UFCS(is_virtual_function)((*cpp2::impl::assert_not_null(n))); }
-#line 364 "reflect.h2"
+#line 363 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_defaultable() const& -> bool { return CPP2_UFCS(is_defaultable_function)((*cpp2::impl::assert_not_null(n))); }
-#line 365 "reflect.h2"
+#line 364 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_constructor() const& -> bool { return CPP2_UFCS(is_constructor)((*cpp2::impl::assert_not_null(n))); }
-#line 366 "reflect.h2"
+#line 365 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_default_constructor() const& -> bool { return CPP2_UFCS(is_default_constructor)((*cpp2::impl::assert_not_null(n))); }
-#line 367 "reflect.h2"
+#line 366 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_move() const& -> bool { return CPP2_UFCS(is_move)((*cpp2::impl::assert_not_null(n))); }
-#line 368 "reflect.h2"
+#line 367 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_swap() const& -> bool { return CPP2_UFCS(is_swap)((*cpp2::impl::assert_not_null(n))); }
-#line 369 "reflect.h2"
+#line 368 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_constructor_with_that() const& -> bool { return CPP2_UFCS(is_constructor_with_that)((*cpp2::impl::assert_not_null(n))); }
-#line 370 "reflect.h2"
+#line 369 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_constructor_with_in_that() const& -> bool { return CPP2_UFCS(is_constructor_with_in_that)((*cpp2::impl::assert_not_null(n))); }
-#line 371 "reflect.h2"
+#line 370 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_constructor_with_move_that() const& -> bool { return CPP2_UFCS(is_constructor_with_move_that)((*cpp2::impl::assert_not_null(n)));  }
-#line 372 "reflect.h2"
+#line 371 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_assignment() const& -> bool { return CPP2_UFCS(is_assignment)((*cpp2::impl::assert_not_null(n))); }
-#line 373 "reflect.h2"
+#line 372 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_assignment_with_that() const& -> bool { return CPP2_UFCS(is_assignment_with_that)((*cpp2::impl::assert_not_null(n))); }
-#line 374 "reflect.h2"
+#line 373 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_assignment_with_in_that() const& -> bool { return CPP2_UFCS(is_assignment_with_in_that)((*cpp2::impl::assert_not_null(n))); }
-#line 375 "reflect.h2"
+#line 374 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_assignment_with_move_that() const& -> bool { return CPP2_UFCS(is_assignment_with_move_that)((*cpp2::impl::assert_not_null(n)));  }
-#line 376 "reflect.h2"
+#line 375 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_destructor() const& -> bool { return CPP2_UFCS(is_destructor)((*cpp2::impl::assert_not_null(n))); }
 
-#line 378 "reflect.h2"
+#line 377 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_copy_or_move() const& -> bool { return is_constructor_with_that() || is_assignment_with_that(); }
 
-#line 380 "reflect.h2"
+#line 379 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_declared_return_type() const& -> bool { return CPP2_UFCS(has_declared_return_type)((*cpp2::impl::assert_not_null(n))); }
-#line 381 "reflect.h2"
+#line 380 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_deduced_return_type() const& -> bool { return CPP2_UFCS(has_deduced_return_type)((*cpp2::impl::assert_not_null(n))); }
-#line 382 "reflect.h2"
+#line 381 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_bool_return_type() const& -> bool { return CPP2_UFCS(has_bool_return_type)((*cpp2::impl::assert_not_null(n))); }
-#line 383 "reflect.h2"
+#line 382 "reflect.h2"
     [[nodiscard]] auto function_declaration::has_non_void_return_type() const& -> bool { return CPP2_UFCS(has_non_void_return_type)((*cpp2::impl::assert_not_null(n))); }
 
-#line 385 "reflect.h2"
+#line 384 "reflect.h2"
     [[nodiscard]] auto function_declaration::unnamed_return_type() const& -> std::string { return CPP2_UFCS(unnamed_return_type_to_string)((*cpp2::impl::assert_not_null(n))); }
 
-#line 387 "reflect.h2"
+#line 386 "reflect.h2"
     [[nodiscard]] auto function_declaration::get_parameters() const& -> std::vector<object_declaration>
 
     {
@@ -1208,19 +1924,19 @@ declaration::declaration(declaration const& that)
         return ret; 
     }
 
-#line 397 "reflect.h2"
+#line 396 "reflect.h2"
     [[nodiscard]] auto function_declaration::is_binary_comparison_function() const& -> bool { return CPP2_UFCS(is_binary_comparison_function)((*cpp2::impl::assert_not_null(n)));  }
 
-#line 399 "reflect.h2"
+#line 398 "reflect.h2"
     auto function_declaration::default_to_virtual() & -> void { static_cast<void>(CPP2_UFCS(make_function_virtual)((*cpp2::impl::assert_not_null(n)))); }
 
-#line 401 "reflect.h2"
+#line 400 "reflect.h2"
     [[nodiscard]] auto function_declaration::make_virtual() & -> bool { return CPP2_UFCS(make_function_virtual)((*cpp2::impl::assert_not_null(n))); }
 
-#line 403 "reflect.h2"
+#line 402 "reflect.h2"
     auto function_declaration::add_initializer(cpp2::impl::in<std::string_view> source) & -> void
 
-#line 406 "reflect.h2"
+#line 405 "reflect.h2"
     {
         if ((*this).is_active() && !(!(has_initializer())) ) { (*this).report_violation(CPP2_CONTRACT_MSG("cannot add an initializer to a function that already has one")); }
         if ((*this).is_active() && !(parent_is_type()) ) { (*this).report_violation(CPP2_CONTRACT_MSG("cannot add an initializer to a function that isn't in a type scope")); }
@@ -1229,7 +1945,7 @@ declaration::declaration(declaration const& that)
         //require( parent_is_type(),
         //         "cannot add an initializer to a function that isn't in a type scope");
 
-#line 412 "reflect.h2"
+#line 411 "reflect.h2"
         auto stmt {parse_statement(source)}; 
         if (!((cpp2::impl::as_<bool>(stmt)))) {
             error("cannot add an initializer that is not a valid statement");
@@ -1242,25 +1958,30 @@ declaration::declaration(declaration const& that)
     function_declaration::function_declaration(function_declaration const& that)
                                 : declaration{ static_cast<declaration const&>(that) }{}
 
-#line 430 "reflect.h2"
+#line 422 "reflect.h2"
+//-----------------------------------------------------------------------
+//  Object declarations
+//
+
+#line 429 "reflect.h2"
     object_declaration::object_declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     )
         : declaration{ n_, s }
-#line 435 "reflect.h2"
+#line 434 "reflect.h2"
     {
 
         if (cpp2::cpp2_default.is_active() && !(CPP2_UFCS(is_object)((*cpp2::impl::assert_not_null(n)))) ) { cpp2::cpp2_default.report_violation(""); }
     }
 
-#line 440 "reflect.h2"
+#line 439 "reflect.h2"
     [[nodiscard]] auto object_declaration::is_const() const& -> bool { return CPP2_UFCS(is_const)((*cpp2::impl::assert_not_null(n))); }
-#line 441 "reflect.h2"
+#line 440 "reflect.h2"
     [[nodiscard]] auto object_declaration::has_wildcard_type() const& -> bool { return CPP2_UFCS(has_wildcard_type)((*cpp2::impl::assert_not_null(n)));  }
 
-#line 443 "reflect.h2"
+#line 442 "reflect.h2"
     [[nodiscard]] auto object_declaration::type() const& -> std::string{
         auto ret {CPP2_UFCS(object_type)((*cpp2::impl::assert_not_null(n)))}; 
         require(!(contains(ret, "(*ERROR*)")), 
@@ -1268,7 +1989,7 @@ declaration::declaration(declaration const& that)
         return ret; 
     }
 
-#line 450 "reflect.h2"
+#line 449 "reflect.h2"
     [[nodiscard]] auto object_declaration::initializer() const& -> std::string{
         auto ret {CPP2_UFCS(object_initializer)((*cpp2::impl::assert_not_null(n)))}; 
         require(!(contains(ret, "(*ERROR*)")), 
@@ -1279,20 +2000,25 @@ declaration::declaration(declaration const& that)
     object_declaration::object_declaration(object_declaration const& that)
                                 : declaration{ static_cast<declaration const&>(that) }{}
 
-#line 466 "reflect.h2"
+#line 458 "reflect.h2"
+//-----------------------------------------------------------------------
+//  Type declarations
+//
+
+#line 465 "reflect.h2"
     type_declaration::type_declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     )
         : declaration{ n_, s }
-#line 471 "reflect.h2"
+#line 470 "reflect.h2"
     {
 
         if (cpp2::cpp2_default.is_active() && !(CPP2_UFCS(is_type)((*cpp2::impl::assert_not_null(n)))) ) { cpp2::cpp2_default.report_violation(""); }
     }
 
-#line 476 "reflect.h2"
+#line 475 "reflect.h2"
     auto type_declaration::reserve_names(cpp2::impl::in<std::string_view> name, auto&& ...etc) const& -> void
     {                           // etc is not declared ':string_view' for compatibility with GCC 10.x
         for ( 
@@ -1307,14 +2033,14 @@ declaration::declaration(declaration const& that)
         }
     }
 
-#line 490 "reflect.h2"
+#line 489 "reflect.h2"
     [[nodiscard]] auto type_declaration::is_polymorphic() const& -> bool { return CPP2_UFCS(is_polymorphic)((*cpp2::impl::assert_not_null(n))); }
-#line 491 "reflect.h2"
+#line 490 "reflect.h2"
     [[nodiscard]] auto type_declaration::is_final() const& -> bool { return CPP2_UFCS(is_type_final)((*cpp2::impl::assert_not_null(n))); }
-#line 492 "reflect.h2"
+#line 491 "reflect.h2"
     [[nodiscard]] auto type_declaration::make_final() & -> bool { return CPP2_UFCS(make_type_final)((*cpp2::impl::assert_not_null(n))); }
 
-#line 494 "reflect.h2"
+#line 493 "reflect.h2"
     [[nodiscard]] auto type_declaration::get_member_functions() const& -> std::vector<function_declaration>
 
     {
@@ -1326,7 +2052,7 @@ declaration::declaration(declaration const& that)
         return ret; 
     }
 
-#line 505 "reflect.h2"
+#line 504 "reflect.h2"
     [[nodiscard]] auto type_declaration::get_member_functions_needing_initializer() const& -> std::vector<function_declaration>
 
     {
@@ -1342,7 +2068,7 @@ declaration::declaration(declaration const& that)
         return ret; 
     }
 
-#line 520 "reflect.h2"
+#line 519 "reflect.h2"
     [[nodiscard]] auto type_declaration::get_member_objects() const& -> std::vector<object_declaration>
 
     {
@@ -1353,7 +2079,7 @@ declaration::declaration(declaration const& that)
         return ret; 
     }
 
-#line 530 "reflect.h2"
+#line 529 "reflect.h2"
     [[nodiscard]] auto type_declaration::get_member_types() const& -> std::vector<type_declaration>
 
     {
@@ -1364,7 +2090,7 @@ declaration::declaration(declaration const& that)
         return ret; 
     }
 
-#line 540 "reflect.h2"
+#line 539 "reflect.h2"
     [[nodiscard]] auto type_declaration::get_member_aliases() const& -> std::vector<alias_declaration>
 
     {
@@ -1375,7 +2101,7 @@ declaration::declaration(declaration const& that)
         return ret; 
     }
 
-#line 550 "reflect.h2"
+#line 549 "reflect.h2"
     [[nodiscard]] auto type_declaration::get_members() const& -> std::vector<declaration>
 
     {
@@ -1386,16 +2112,16 @@ declaration::declaration(declaration const& that)
         return ret; 
     }
 
-#line 560 "reflect.h2"
+#line 559 "reflect.h2"
     [[nodiscard]] auto type_declaration::query_declared_value_set_functions() const& -> query_declared_value_set_functions_ret
 
-#line 567 "reflect.h2"
+#line 566 "reflect.h2"
     {
             cpp2::impl::deferred_init<bool> out_this_in_that;
             cpp2::impl::deferred_init<bool> out_this_move_that;
             cpp2::impl::deferred_init<bool> inout_this_in_that;
             cpp2::impl::deferred_init<bool> inout_this_move_that;
-#line 568 "reflect.h2"
+#line 567 "reflect.h2"
         auto declared {CPP2_UFCS(find_declared_value_set_functions)((*cpp2::impl::assert_not_null(n)))}; 
         out_this_in_that.construct(declared.out_this_in_that != nullptr);
         out_this_move_that.construct(declared.out_this_move_that != nullptr);
@@ -1403,7 +2129,7 @@ declaration::declaration(declaration const& that)
         inout_this_move_that.construct(cpp2::move(declared).inout_this_move_that != nullptr);
     return  { std::move(out_this_in_that.value()), std::move(out_this_move_that.value()), std::move(inout_this_in_that.value()), std::move(inout_this_move_that.value()) }; }
 
-#line 575 "reflect.h2"
+#line 574 "reflect.h2"
     auto type_declaration::add_member(cpp2::impl::in<std::string_view> source) & -> void
     {
         auto decl {parse_statement(source)}; 
@@ -1418,25 +2144,30 @@ declaration::declaration(declaration const& that)
                  std::string("unexpected error while attempting to add member:\n") + source);
     }
 
-#line 589 "reflect.h2"
+#line 588 "reflect.h2"
     auto type_declaration::remove_marked_members() & -> void { CPP2_UFCS(type_remove_marked_members)((*cpp2::impl::assert_not_null(n)));  }
-#line 590 "reflect.h2"
+#line 589 "reflect.h2"
     auto type_declaration::remove_all_members() & -> void { CPP2_UFCS(type_remove_all_members)((*cpp2::impl::assert_not_null(n))); }
 
-#line 592 "reflect.h2"
+#line 591 "reflect.h2"
     auto type_declaration::disable_member_function_generation() & -> void { CPP2_UFCS(type_disable_member_function_generation)((*cpp2::impl::assert_not_null(n)));  }
 
     type_declaration::type_declaration(type_declaration const& that)
                                 : declaration{ static_cast<declaration const&>(that) }{}
 
-#line 603 "reflect.h2"
+#line 595 "reflect.h2"
+//-----------------------------------------------------------------------
+//  Alias declarations
+//
+
+#line 602 "reflect.h2"
     alias_declaration::alias_declaration(
 
         declaration_node* n_, 
         cpp2::impl::in<compiler_services> s
     )
         : declaration{ n_, s }
-#line 608 "reflect.h2"
+#line 607 "reflect.h2"
     {
 
         if (cpp2::cpp2_default.is_active() && !(CPP2_UFCS(is_alias)((*cpp2::impl::assert_not_null(n)))) ) { cpp2::cpp2_default.report_violation(""); }
@@ -1445,13 +2176,39 @@ declaration::declaration(declaration const& that)
     alias_declaration::alias_declaration(alias_declaration const& that)
                                 : declaration{ static_cast<declaration const&>(that) }{}
 
-#line 627 "reflect.h2"
+#line 614 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//  Metafunctions - these are hardwired for now until we get to the
+//  step of writing a Cpp2 interpreter to run inside the compiler
+//
+//-----------------------------------------------------------------------
+//
+
+//-----------------------------------------------------------------------
+//  Some common metafunction helpers (metafunctions are just functions,
+//  so they can be factored as usual)
+//
+#line 626 "reflect.h2"
 auto add_virtual_destructor(meta::type_declaration& t) -> void
 {
     CPP2_UFCS(add_member)(t, "operator=: (virtual move this) = { }");
 }
 
-#line 645 "reflect.h2"
+#line 632 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//      "... an abstract base class defines an interface ..."
+//
+//          -- Stroustrup (The Design and Evolution of C++, 12.3.1)
+//
+//-----------------------------------------------------------------------
+//
+//  interface
+//
+//  an abstract base class having only pure virtual functions
+//
+#line 644 "reflect.h2"
 auto interface(meta::type_declaration& t) -> void
 {
     auto has_dtor {false}; 
@@ -1478,7 +2235,27 @@ auto interface(meta::type_declaration& t) -> void
     }
 }
 
-#line 691 "reflect.h2"
+#line 671 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     "C.35: A base class destructor should be either public and
+//      virtual, or protected and non-virtual."
+//
+//     "[C.43] ... a base class should not be copyable, and so does not
+//      necessarily need a default constructor."
+//
+//          -- Stroustrup, Sutter, et al. (C++ Core Guidelines)
+//
+//-----------------------------------------------------------------------
+//
+//  polymorphic_base
+//
+//  A pure polymorphic base type that is not copyable, and whose
+//  destructor is either public and virtual or protected and nonvirtual.
+//
+//  Unlike an interface, it can have nonpublic and nonvirtual functions.
+//
+#line 690 "reflect.h2"
 auto polymorphic_base(meta::type_declaration& t) -> void
 {
     auto has_dtor {false}; 
@@ -1503,10 +2280,31 @@ auto polymorphic_base(meta::type_declaration& t) -> void
     }
 }
 
-#line 736 "reflect.h2"
+#line 715 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     "... A totally ordered type ... requires operator<=> that
+//      returns std::strong_ordering. If the function is not
+//      user-written, a lexicographical memberwise implementation
+//      is generated by default..."
+//
+//          -- P0707R4, section 3
+//
+//      Note: This feature derived from Cpp2 was already adopted
+//            into Standard C++ via paper P0515, so most of the
+//            heavy lifting is done by the Cpp1 C++20/23 compiler,
+//            including the memberwise default semantics
+//            (In contrast, cppfront has to do the work itself for
+//            default memberwise semantics for operator= assignment
+//            as those aren't yet part of Standard C++)
+//
+//-----------------------------------------------------------------------
+//
+
+#line 735 "reflect.h2"
 auto ordered_impl(
     meta::type_declaration& t, 
-    cpp2::impl::in<std::string_view> ordering
+    cpp2::impl::in<std::string_view> ordering// must be "strong_ordering" etc.
 ) -> void
 {
     auto has_spaceship {false}; 
@@ -1528,25 +2326,53 @@ auto ordered_impl(
     }
 }
 
-#line 765 "reflect.h2"
+//-----------------------------------------------------------------------
+//  ordered - a totally ordered type
+//
+//  Note: the ordering that should be encouraged as default gets the nice name
+//
+#line 764 "reflect.h2"
 auto ordered(meta::type_declaration& t) -> void
 {
     ordered_impl(t, "strong_ordering");
 }
 
-#line 773 "reflect.h2"
+//-----------------------------------------------------------------------
+//  weakly_ordered - a weakly ordered type
+//
+#line 772 "reflect.h2"
 auto weakly_ordered(meta::type_declaration& t) -> void
 {
     ordered_impl(t, "weak_ordering");
 }
 
-#line 781 "reflect.h2"
+//-----------------------------------------------------------------------
+//  partially_ordered - a partially ordered type
+//
+#line 780 "reflect.h2"
 auto partially_ordered(meta::type_declaration& t) -> void
 {
     ordered_impl(t, "partial_ordering");
 }
 
-#line 803 "reflect.h2"
+#line 786 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     "A value is ... a regular type. It must have all public
+//      default construction, copy/move construction/assignment,
+//      and destruction, all of which are generated by default
+//      if not user-written; and it must not have any protected
+//      or virtual functions (including the destructor)."
+//
+//          -- P0707R4, section 3
+//
+//-----------------------------------------------------------------------
+//
+//  copyable
+//
+//  A type with (copy and move) x (construction and assignment)
+//
+#line 802 "reflect.h2"
 auto copyable(meta::type_declaration& t) -> void
 {
     //  If the user explicitly wrote any of the copy/move functions,
@@ -1574,7 +2400,14 @@ auto copyable(meta::type_declaration& t) -> void
     }}
 }
 
-#line 837 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//  basic_value
+//
+//  A regular type: copyable, plus has public default construction
+//  and no protected or virtual functions
+//
+#line 836 "reflect.h2"
 auto basic_value(meta::type_declaration& t) -> void
 {
     CPP2_UFCS(copyable)(t);
@@ -1593,28 +2426,61 @@ auto basic_value(meta::type_declaration& t) -> void
     }
 }
 
-#line 865 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     "A 'value' is a totally ordered basic_value..."
+//
+//          -- P0707R4, section 3
+//
+//  value - a value type that is totally ordered
+//
+//  Note: the ordering that should be encouraged as default gets the nice name
+//
+#line 864 "reflect.h2"
 auto value(meta::type_declaration& t) -> void
 {
     CPP2_UFCS(ordered)(t);
     CPP2_UFCS(basic_value)(t);
 }
 
-#line 871 "reflect.h2"
+#line 870 "reflect.h2"
 auto weakly_ordered_value(meta::type_declaration& t) -> void
 {
     CPP2_UFCS(weakly_ordered)(t);
     CPP2_UFCS(basic_value)(t);
 }
 
-#line 877 "reflect.h2"
+#line 876 "reflect.h2"
 auto partially_ordered_value(meta::type_declaration& t) -> void
 {
     CPP2_UFCS(partially_ordered)(t);
     CPP2_UFCS(basic_value)(t);
 }
 
-#line 906 "reflect.h2"
+#line 883 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     C.20: If you can avoid defining default operations, do
+//
+//     ##### Reason
+//
+//     It's the simplest and gives the cleanest semantics.
+//
+//     ...
+//
+//     This is known as "the rule of zero".
+//
+//          -- C++ Core Guidelines
+//             C.20: If you can avoid defining any default operations, do
+//             <https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rc-zero>
+//
+//-----------------------------------------------------------------------
+//
+//  cpp1_rule_of_zero
+//
+//  a type without declared copy/move/destructor functions
+//
+#line 905 "reflect.h2"
 auto cpp1_rule_of_zero(meta::type_declaration& t) -> void
 {
     for ( auto& mf : CPP2_UFCS(get_member_functions)(t) ) 
@@ -1627,7 +2493,32 @@ auto cpp1_rule_of_zero(meta::type_declaration& t) -> void
     CPP2_UFCS(disable_member_function_generation)(t);
 }
 
-#line 943 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     "By definition, a `struct` is a `class` in which members
+//      are by default `public`; that is,
+//
+//      	struct s { ...
+//
+//      is simply shorthand for
+//
+//          class s { public: ...
+//
+//      ... Which style you use depends on circumstances and taste.
+//      I usually prefer to use `struct` for classes that have all
+//      data `public`."
+//
+//          -- Stroustrup (The C++ Programming Language, 3rd ed., p. 234)
+//
+//-----------------------------------------------------------------------
+//
+//  struct
+//
+//  a type with only public bases, objects, and functions,
+//  no virtual functions, and no user-defined constructors
+//  (i.e., no invariants) or assignment or destructors.
+//
+#line 942 "reflect.h2"
 auto cpp2_struct(meta::type_declaration& t) -> void
 {
     std::string ctor_params {}; 
@@ -1679,7 +2570,26 @@ value_member_info::value_member_info(auto const& name_, auto const& type_, auto 
                                                                , type{ type_ }
                                                                , value{ value_ }{}
 
-#line 1013 "reflect.h2"
+#line 989 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     "C enumerations constitute a curiously half-baked concept. ...
+//      the cleanest way out was to deem each enumeration a separate type."
+//
+//          -- Stroustrup (The Design and Evolution of C++, 11.7)
+//
+//     "An enumeration is a distinct type ... with named constants"
+//
+//          -- ISO C++ Standard
+//
+//-----------------------------------------------------------------------
+//
+//  basic_enum
+//
+//  a type together with named constants that are its possible values
+//
+
+#line 1012 "reflect.h2"
 auto basic_enum(
     meta::type_declaration& t, 
     auto const& nextval, 
@@ -1704,7 +2614,7 @@ auto basic_enum(
 {
 std::string value{"-1"};
 
-#line 1036 "reflect.h2"
+#line 1035 "reflect.h2"
     for ( 
           auto const& m : CPP2_UFCS(get_members)(t) ) 
     if (  CPP2_UFCS(is_member_object)(m)) 
@@ -1746,7 +2656,7 @@ std::string value{"-1"};
     }
 }
 
-#line 1076 "reflect.h2"
+#line 1075 "reflect.h2"
     if ((CPP2_UFCS(empty)(enumerators))) {
         CPP2_UFCS(error)(t, "an enumeration must contain at least one enumerator value");
         return ; 
@@ -1797,7 +2707,7 @@ std::string value{"-1"};
         }
     }
 
-#line 1127 "reflect.h2"
+#line 1126 "reflect.h2"
     //  2. Replace: Erase the contents and replace with modified contents
     //
     //  Note that most values and functions are declared as '==' compile-time values, i.e. Cpp1 'constexpr'
@@ -1847,7 +2757,7 @@ std::string to_string_impl{"    to_string_impl: (this, prefix: std::string_view"
     //  Provide 'to_string' and 'to_code' functions to print enumerator
     //  name(s) as human-readable strings or as code expressions
 
-#line 1174 "reflect.h2"
+#line 1173 "reflect.h2"
     {
         if (bitwise) {
             to_string_impl += ", separator: std::string_view ) -> std::string = { \n"
@@ -1888,7 +2798,7 @@ std::string to_string_impl{"    to_string_impl: (this, prefix: std::string_view"
     }
 }
 
-#line 1213 "reflect.h2"
+#line 1212 "reflect.h2"
     if (bitwise) {
         CPP2_UFCS(add_member)(t, "    to_string: (this) -> std::string = to_string_impl( \"\", \", \" );");
         CPP2_UFCS(add_member)(t, "    to_code  : (this) -> std::string = to_string_impl( \"" + cpp2::to_string(CPP2_UFCS(name)(t)) + "::\", \" | \" );");
@@ -1902,7 +2812,7 @@ std::string from_string{"    from_string: (s: std::string_view) -> " + cpp2::to_
 
     //  Provide a 'from_string' function to parse strings into enumerators    
 
-#line 1224 "reflect.h2"
+#line 1223 "reflect.h2"
     {
         std::string_view prefix {""}; 
         std::string_view combine_op {"return"}; 
@@ -1924,7 +2834,7 @@ std::string from_string{"    from_string: (s: std::string_view) -> " + cpp2::to_
 {
 std::string_view else_{""};
 
-#line 1244 "reflect.h2"
+#line 1243 "reflect.h2"
         for ( 
               auto const& e : cpp2::move(enumerators) ) {
             from_string += "            " + cpp2::to_string(else_) + "if \"" + cpp2::to_string(e.name) + "\" == x { " + cpp2::to_string(combine_op) + " " + cpp2::to_string(e.name) + "; }\n";
@@ -1932,7 +2842,7 @@ std::string_view else_{""};
         }
 }
 
-#line 1250 "reflect.h2"
+#line 1249 "reflect.h2"
         if (bitwise) {
             from_string += "            else { break outer; }\n"
                            "        }\n"
@@ -1948,11 +2858,21 @@ std::string_view else_{""};
     }
 }
 
-#line 1264 "reflect.h2"
+#line 1263 "reflect.h2"
     CPP2_UFCS(add_member)(t, "    from_code: (s: std::string_view) -> " + cpp2::to_string(CPP2_UFCS(name)(t)) + " = { str: std::string = s; return from_string( cpp2::string_util::replace_all(str, \"" + cpp2::to_string(CPP2_UFCS(name)(t)) + "::\", \"\" ) ); }");
 }
 
-#line 1277 "reflect.h2"
+#line 1267 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//    "An enum[...] is a totally ordered value type that stores a
+//     value of its enumerators's type, and otherwise has only public
+//     member variables of its enumerator's type, all of which are
+//     naturally scoped because they are members of a type."
+//
+//          -- P0707R4, section 3
+//
+#line 1276 "reflect.h2"
 auto cpp2_enum(meta::type_declaration& t) -> void
 {
     //  Let basic_enum do its thing, with an incrementing value generator
@@ -1969,7 +2889,18 @@ auto cpp2_enum(meta::type_declaration& t) -> void
     );
 }
 
-#line 1304 "reflect.h2"
+#line 1293 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     "flag_enum expresses an enumeration that stores values
+//      corresponding to bitwise-or'd enumerators. The enumerators must
+//      be powers of two, and are automatically generated [...] A none
+//      value is provided [...] Operators | and & are provided to
+//      combine and extract values."
+//
+//          -- P0707R4, section 3
+//
+#line 1303 "reflect.h2"
 auto flag_enum(meta::type_declaration& t) -> void
 {
     //  Let basic_enum do its thing, with a power-of-two value generator
@@ -1991,7 +2922,32 @@ auto flag_enum(meta::type_declaration& t) -> void
     );
 }
 
-#line 1350 "reflect.h2"
+#line 1325 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//     "As with void*, programmers should know that unions [...] are
+//      inherently dangerous, should be avoided wherever possible,
+//      and should be handled with special care when actually needed."
+//
+//          -- Stroustrup (The Design and Evolution of C++, 14.3.4.1)
+//
+//     "C++17 needs a type-safe union... The implications of the
+//      consensus `variant` design are well understood and have been
+//      explored over several LEWG discussions, over a thousand emails,
+//      a joint LEWG/EWG session, and not to mention 12 years of
+//      experience with Boost and other libraries."
+//
+//          -- Axel Naumann, in P0088 (wg21.link/p0088),
+//             the adopted proposal for C++17 std::variant
+//
+//-----------------------------------------------------------------------
+//
+//  union
+//
+//  a type that contains exactly one of a fixed set of values at a time
+//
+
+#line 1349 "reflect.h2"
 auto cpp2_union(meta::type_declaration& t) -> void
 {
     std::vector<value_member_info> alternatives {}; 
@@ -2000,7 +2956,7 @@ auto value{0};
 
     //  1. Gather: All the user-written members, and find/compute the max size
 
-#line 1357 "reflect.h2"
+#line 1356 "reflect.h2"
     for ( 
 
            auto const& m : CPP2_UFCS(get_members)(t) )  { do 
@@ -2030,7 +2986,7 @@ auto value{0};
     } while (false); ++value; }
 }
 
-#line 1385 "reflect.h2"
+#line 1384 "reflect.h2"
     std::string discriminator_type {}; 
     if (cpp2::impl::cmp_less(CPP2_UFCS(ssize)(alternatives),std::numeric_limits<cpp2::i8>::max())) {
         discriminator_type = "i8";
@@ -2045,7 +3001,7 @@ auto value{0};
         discriminator_type = "i64";
     }}}
 
-#line 1400 "reflect.h2"
+#line 1399 "reflect.h2"
     //  2. Replace: Erase the contents and replace with modified contents
 
     CPP2_UFCS(remove_marked_members)(t);
@@ -2054,7 +3010,7 @@ std::string storage{"    _storage: cpp2::aligned_storage<cpp2::max( "};
 
     //  Provide storage
 
-#line 1406 "reflect.h2"
+#line 1405 "reflect.h2"
     {
         for ( 
               auto const& e : alternatives ) {
@@ -2074,7 +3030,7 @@ std::string storage{"    _storage: cpp2::aligned_storage<cpp2::max( "};
 }
 
     //  Provide discriminator
-#line 1424 "reflect.h2"
+#line 1423 "reflect.h2"
     CPP2_UFCS(add_member)(t, "    _discriminator: " + cpp2::to_string(cpp2::move(discriminator_type)) + " = -1;\n");
 
     //  Add the alternatives: is_alternative, get_alternative, and set_alternative
@@ -2110,7 +3066,7 @@ std::string destroy{"    private _destroy: (inout this) = {\n"};
 
     //  Add destroy
 
-#line 1457 "reflect.h2"
+#line 1456 "reflect.h2"
     {
         for ( 
               auto const& a : alternatives ) {
@@ -2124,7 +3080,7 @@ std::string destroy{"    private _destroy: (inout this) = {\n"};
 }
 
     //  Add the destructor
-#line 1469 "reflect.h2"
+#line 1468 "reflect.h2"
     CPP2_UFCS(add_member)(t, "    operator=: (move this) = { _destroy(); _ = this; }");
 
     //  Add default constructor
@@ -2134,7 +3090,7 @@ std::string value_set{""};
 
     //  Add copy/move construction and assignment
 
-#line 1476 "reflect.h2"
+#line 1475 "reflect.h2"
     {
         for ( 
               auto const& a : cpp2::move(alternatives) ) {
@@ -2154,16 +3110,2081 @@ std::string value_set{""};
                     );
     }
 }
-#line 1494 "reflect.h2"
+#line 1493 "reflect.h2"
 }
 
-#line 1501 "reflect.h2"
+#line 1496 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//  print - output a pretty-printed visualization of t
+//
+#line 1500 "reflect.h2"
 auto print(cpp2::impl::in<meta::type_declaration> t) -> void
 {
     std::cout << CPP2_UFCS(print)(t) << "\n";
 }
 
-#line 1521 "reflect.h2"
+} // end meta subnamespace
+
+#line 1508 "reflect.h2"
+//-----------------------------------------------------------------------
+//
+//  regex - creates regular expressions from members
+//
+//  Each member that starts with `regex` is replaced by a regular expression
+//  of the initializer string. E.g.:
+//  ```
+//    regex := "ab";
+//  ```
+//  is replaced with
+//  ```
+//    regex := ::cpp2::regex::regular_expression<...>;
+//  ```
+//
+namespace regex {
+
+
+constexpr expression_flags::expression_flags(cpp2::impl::in<cpp2::i64> _val)
+                                                          : _value{ cpp2::unsafe_narrow<cpp2::u8>(_val) } {  }
+
+constexpr auto expression_flags::operator=(cpp2::impl::in<cpp2::i64> _val) -> expression_flags&  { 
+                                                          _value = cpp2::unsafe_narrow<cpp2::u8>(_val);
+                                                          return *this; }
+constexpr auto expression_flags::operator|=(expression_flags const& that) & -> void { _value |= that._value; }
+constexpr auto expression_flags::operator&=(expression_flags const& that) & -> void { _value &= that._value; }
+constexpr auto expression_flags::operator^=(expression_flags const& that) & -> void { _value ^= that._value; }
+[[nodiscard]] constexpr auto expression_flags::operator|(expression_flags const& that) const& -> expression_flags { return _value | that._value; }
+[[nodiscard]] constexpr auto expression_flags::operator&(expression_flags const& that) const& -> expression_flags { return _value & that._value; }
+[[nodiscard]] constexpr auto expression_flags::operator^(expression_flags const& that) const& -> expression_flags { return _value ^ that._value; }
+[[nodiscard]] constexpr auto expression_flags::has(expression_flags const& that) const& -> bool { return _value & that._value; }
+constexpr auto expression_flags::set(expression_flags const& that) & -> void { _value |= that._value; }
+constexpr auto expression_flags::clear(expression_flags const& that) & -> void { _value &= ~that._value; }
+inline CPP2_CONSTEXPR expression_flags expression_flags::case_insensitive{ 1 };
+
+inline CPP2_CONSTEXPR expression_flags expression_flags::multiple_lines{ 2 };
+
+inline CPP2_CONSTEXPR expression_flags expression_flags::single_line{ 4 };
+
+inline CPP2_CONSTEXPR expression_flags expression_flags::no_group_captures{ 8 };
+
+inline CPP2_CONSTEXPR expression_flags expression_flags::perl_code_syntax{ 16 };
+
+inline CPP2_CONSTEXPR expression_flags expression_flags::perl_code_syntax_in_classes{ 32 };
+
+inline CPP2_CONSTEXPR expression_flags expression_flags::none{ 0 };
+
+[[nodiscard]] constexpr auto expression_flags::get_raw_value() const& -> cpp2::u8 { return _value; }
+constexpr expression_flags::expression_flags()
+                                        : _value{ none._value }{}
+constexpr expression_flags::expression_flags(expression_flags const& that)
+                                              : _value{ that._value }{}
+constexpr auto expression_flags::operator=(expression_flags const& that) -> expression_flags& {
+                                              _value = that._value;
+                                              return *this;}
+constexpr expression_flags::expression_flags(expression_flags&& that) noexcept
+                                              : _value{ std::move(that)._value }{}
+constexpr auto expression_flags::operator=(expression_flags&& that) noexcept -> expression_flags& {
+                                              _value = std::move(that)._value;
+                                              return *this;}
+[[nodiscard]] auto expression_flags::to_string_impl(cpp2::impl::in<std::string_view> prefix, cpp2::impl::in<std::string_view> separator) const& -> std::string{
+
+std::string ret {"("}; 
+
+std::string sep {}; 
+if ((*this) == none) {return "(none)"; }
+
+auto pref {cpp2::to_string(prefix)}; 
+if (((*this) & case_insensitive) == case_insensitive) {ret += sep + pref + "case_insensitive";sep = separator;}
+if (((*this) & multiple_lines) == multiple_lines) {ret += sep + pref + "multiple_lines";sep = separator;}
+if (((*this) & single_line) == single_line) {ret += sep + pref + "single_line";sep = separator;}
+if (((*this) & no_group_captures) == no_group_captures) {ret += sep + pref + "no_group_captures";sep = separator;}
+if (((*this) & perl_code_syntax) == perl_code_syntax) {ret += sep + pref + "perl_code_syntax";sep = separator;}
+if (((*this) & perl_code_syntax_in_classes) == perl_code_syntax_in_classes) {ret += sep + cpp2::move(pref) + "perl_code_syntax_in_classes";sep = separator;}
+return cpp2::move(ret) + ")"; 
+}
+
+[[nodiscard]] auto expression_flags::to_string() const& -> std::string { return to_string_impl("", ", "); }
+[[nodiscard]] auto expression_flags::to_code() const& -> std::string { return to_string_impl("expression_flags::", " | "); }
+[[nodiscard]] auto expression_flags::from_string(cpp2::impl::in<std::string_view> s) -> expression_flags{
+
+auto ret {none}; 
+do {{
+for ( auto const& x : cpp2::string_util::split_string_list(s) ) {
+if ("case_insensitive" == x) {ret |= case_insensitive;}
+else {if ("multiple_lines" == x) {ret |= multiple_lines;}
+else {if ("single_line" == x) {ret |= single_line;}
+else {if ("no_group_captures" == x) {ret |= no_group_captures;}
+else {if ("perl_code_syntax" == x) {ret |= perl_code_syntax;}
+else {if ("perl_code_syntax_in_classes" == x) {ret |= perl_code_syntax_in_classes;}
+else {if ("none" == x) {ret |= none;}
+else {goto BREAK_outer;}
+#line 1 "reflect.h2"
+}}}}}}
+}
+
+return ret; 
+} CPP2_CONTINUE_BREAK(outer) }
+ while ( 
+false
+);
+CPP2_UFCS(report_violation)(cpp2::type_safety, CPP2_UFCS(c_str)(("can't convert string '" + cpp2::to_string(s) + "' to flag_enum of type expression_flags")));
+return none; 
+}
+
+[[nodiscard]] auto expression_flags::from_code(cpp2::impl::in<std::string_view> s) -> expression_flags{
+std::string str {s}; return from_string(cpp2::string_util::replace_all(cpp2::move(str), "expression_flags::", "")); }
+
+#line 1526 "reflect.h2"
+//  Possible modifiers for a regular expression.
+//
+
+#line 1530 "reflect.h2"
+                                  // mod: i
+                                  // mod: m
+                                  // mod: s
+                                  // mod: n
+                                  // mod: x
+                                  // mod: xx
+
+#line 1539 "reflect.h2"
+//  Tokens for regular expressions.
+//
+
+// Basic class for a regex token.
+//
+
+#line 1548 "reflect.h2"
+    regex_token::regex_token(cpp2::impl::in<std::string> str)
+        : string_rep{ str }{
+
+#line 1550 "reflect.h2"
+    }
+
+#line 1552 "reflect.h2"
+    regex_token::regex_token()
+        : string_rep{ "" }{
+
+#line 1554 "reflect.h2"
+    }
+
+    //parse: (inout ctx: parse_context) -> token_ptr;
+                                                                       // Generate the matching code.
+
+#line 1559 "reflect.h2"
+    auto regex_token::add_groups([[maybe_unused]] std::set<int>& unnamed_param_2) const -> void{}// Adds all group indices to the set.
+#line 1560 "reflect.h2"
+    [[nodiscard]] auto regex_token::to_string() const& -> std::string{return string_rep; }// Create a string representation.
+#line 1561 "reflect.h2"
+    auto regex_token::set_string(cpp2::impl::in<std::string> s) & -> void{string_rep = s; }
+
+    regex_token::~regex_token() noexcept{}// Set the string representation.
+
+#line 1576 "reflect.h2"
+    regex_token_check::regex_token_check(cpp2::impl::in<std::string> str, cpp2::impl::in<std::string> check_)
+        : regex_token{ str }
+        , check{ check_ }{
+
+#line 1579 "reflect.h2"
+    }
+
+#line 1581 "reflect.h2"
+    auto regex_token_check::generate_code(generation_context& ctx) const -> void{
+        ctx.add_check(check + "(" + ctx.match_parameters() + ")");
+    }
+
+    regex_token_check::~regex_token_check() noexcept{}
+
+#line 1595 "reflect.h2"
+    regex_token_code::regex_token_code(cpp2::impl::in<std::string> str, cpp2::impl::in<std::string> code_)
+        : regex_token{ str }
+        , code{ code_ }{
+
+#line 1598 "reflect.h2"
+    }
+
+#line 1600 "reflect.h2"
+    auto regex_token_code::generate_code(generation_context& ctx) const -> void{
+        ctx.add(code);
+    }
+
+    regex_token_code::~regex_token_code() noexcept{}
+
+#line 1612 "reflect.h2"
+    regex_token_empty::regex_token_empty(cpp2::impl::in<std::string> str)
+        : regex_token{ str }{
+
+#line 1614 "reflect.h2"
+    }
+
+#line 1616 "reflect.h2"
+    auto regex_token_empty::generate_code([[maybe_unused]] generation_context& unnamed_param_2) const -> void{
+        // Nothing.
+    }
+
+    regex_token_empty::~regex_token_empty() noexcept{}
+
+#line 1630 "reflect.h2"
+    regex_token_list::regex_token_list(cpp2::impl::in<token_vec> t)
+        : regex_token{ gen_string(t) }
+        , tokens{ t }{
+
+#line 1633 "reflect.h2"
+    }
+
+#line 1635 "reflect.h2"
+    auto regex_token_list::generate_code(generation_context& ctx) const -> void{
+        for ( auto const& token : tokens ) {
+            (*cpp2::impl::assert_not_null(token)).generate_code(ctx);
+        }
+    }
+
+#line 1641 "reflect.h2"
+    auto regex_token_list::add_groups(std::set<int>& groups) const -> void{
+        for ( auto const& token : tokens ) {
+            (*cpp2::impl::assert_not_null(token)).add_groups(groups);
+        }
+    }
+
+#line 1647 "reflect.h2"
+    [[nodiscard]] auto regex_token_list::gen_string(cpp2::impl::in<token_vec> vec) -> std::string{
+        std::string r {""}; 
+        for ( auto const& token : vec ) {
+            r += (*cpp2::impl::assert_not_null(token)).to_string();
+        }
+        return r; 
+    }
+
+    regex_token_list::~regex_token_list() noexcept{}
+
+#line 1670 "reflect.h2"
+    auto parse_context_group_state::next_alternative() & -> void{
+        token_vec new_list {}; 
+        std::swap(new_list, cur_match_list);
+        post_process_list(new_list);
+        static_cast<void>(alternate_match_lists.insert(alternate_match_lists.end(), CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_list>)(cpp2::shared, cpp2::move(new_list))));
+    }
+
+#line 1678 "reflect.h2"
+    auto parse_context_group_state::swap(parse_context_group_state& t) & -> void{
+        std::swap(cur_match_list, t.cur_match_list);
+        std::swap(alternate_match_lists, t.alternate_match_lists);
+        std::swap(modifiers, t.modifiers);
+    }
+
+#line 1685 "reflect.h2"
+    [[nodiscard]] auto parse_context_group_state::get_as_token() & -> token_ptr{
+        if (alternate_match_lists.empty()) {
+            post_process_list(cur_match_list);
+            return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_list>)(cpp2::shared, cur_match_list); 
+        }
+        else {
+            next_alternative();
+            return CPP2_UFCS_TEMPLATE(cpp2_new<alternative_token_gen>)(cpp2::shared, alternate_match_lists); 
+        }
+    }
+
+#line 1697 "reflect.h2"
+    auto parse_context_group_state::add(cpp2::impl::in<token_ptr> token) & -> void{
+        cur_match_list.push_back(token);
+    }
+
+#line 1702 "reflect.h2"
+    [[nodiscard]] auto parse_context_group_state::empty() const& -> bool { return cur_match_list.empty();  }
+
+#line 1706 "reflect.h2"
+    auto parse_context_group_state::post_process_list(token_vec& list) -> void{
+        // Merge all characters
+        auto merge_pos {list.begin()}; 
+        for( ; merge_pos != list.end(); (++merge_pos) ) {
+            if (cpp2::impl::is<char_token>(*cpp2::impl::assert_not_null(*cpp2::impl::assert_not_null(merge_pos)))) {
+                auto combine_pos {merge_pos + 1}; 
+                while( combine_pos != list.end() && cpp2::impl::is<char_token>(*cpp2::impl::assert_not_null(*cpp2::impl::assert_not_null(combine_pos))) ) {// The erase advances combine_pos
+                    (cpp2::impl::as_<char_token>(*cpp2::impl::assert_not_null(*cpp2::impl::assert_not_null(merge_pos)))).append(cpp2::impl::as_<char_token>(*cpp2::impl::assert_not_null(*cpp2::impl::assert_not_null(combine_pos))));
+                    combine_pos = list.erase(combine_pos);
+                }
+            }
+        }
+    }
+
+    parse_context_group_state::parse_context_group_state(auto const& cur_match_list_, auto const& alternate_match_lists_, auto const& modifiers_)
+                                                                                              : cur_match_list{ cur_match_list_ }
+                                                                                              , alternate_match_lists{ alternate_match_lists_ }
+                                                                                              , modifiers{ modifiers_ }{}
+parse_context_group_state::parse_context_group_state(){}
+
+#line 1732 "reflect.h2"
+    [[nodiscard]] auto parse_context_branch_reset_state::next() & -> int{
+        auto g {cur_group}; 
+        cur_group += 1;
+        max_group = max(max_group, cur_group);
+
+        return g; 
+    }
+
+#line 1741 "reflect.h2"
+    auto parse_context_branch_reset_state::set_next(cpp2::impl::in<int> g) & -> void{
+        cur_group = g;
+        max_group = max(max_group, g);
+    }
+
+#line 1747 "reflect.h2"
+    auto parse_context_branch_reset_state::next_alternative() & -> void{
+        if (is_active) {
+            cur_group = from;
+        }
+    }
+
+#line 1754 "reflect.h2"
+    auto parse_context_branch_reset_state::set_active_reset(cpp2::impl::in<int> restart) & -> void{
+        is_active = true;
+        cur_group = restart;
+        from      = restart;
+        max_group = restart;
+    }
+
+    parse_context_branch_reset_state::parse_context_branch_reset_state(auto const& is_active_, auto const& cur_group_, auto const& max_group_, auto const& from_)
+                                                                                    : is_active{ is_active_ }
+                                                                                    , cur_group{ cur_group_ }
+                                                                                    , max_group{ max_group_ }
+                                                                                    , from{ from_ }{}
+parse_context_branch_reset_state::parse_context_branch_reset_state(){}
+
+#line 1782 "reflect.h2"
+    parse_context::parse_context(cpp2::impl::in<std::string_view> r, auto const& e)
+        : regex{ r }
+        , root{ CPP2_UFCS_TEMPLATE_NONLOCAL(cpp2_new<regex_token_empty>)(cpp2::shared, "") }
+        , error_out{ e }{
+
+#line 1786 "reflect.h2"
+    }
+
+#line 1792 "reflect.h2"
+    [[nodiscard]] auto parse_context::start_group() & -> parse_context_group_state
+    {
+        parse_context_group_state old_state {}; 
+        old_state.swap(cur_group_state);
+        cur_group_state.modifiers = old_state.modifiers;
+
+        return old_state; 
+    }
+
+#line 1802 "reflect.h2"
+    [[nodiscard]] auto parse_context::end_group(cpp2::impl::in<parse_context_group_state> old_state) & -> token_ptr
+    {
+        auto inner {cur_group_state.get_as_token()}; 
+        cur_group_state = old_state;
+        return inner; 
+    }
+
+#line 1809 "reflect.h2"
+    [[nodiscard]] auto parse_context::get_modifiers() const& -> expression_flags{
+        return cur_group_state.modifiers; 
+    }
+
+#line 1813 "reflect.h2"
+    auto parse_context::set_modifiers(cpp2::impl::in<expression_flags> mod) & -> void{
+        cur_group_state.modifiers = mod;
+    }
+
+#line 1820 "reflect.h2"
+    [[nodiscard]] auto parse_context::branch_reset_new_state() & -> parse_context_branch_reset_state
+    {
+        parse_context_branch_reset_state old_state {}; 
+        std::swap(old_state, cur_branch_reset_state);
+
+        cur_branch_reset_state.set_active_reset(old_state.cur_group);
+        return old_state; 
+    }
+
+#line 1829 "reflect.h2"
+    auto parse_context::branch_reset_restore_state(cpp2::impl::in<parse_context_branch_reset_state> old_state) & -> void
+    {
+        auto max_group {cur_branch_reset_state.max_group}; 
+        cur_branch_reset_state = old_state;
+        cur_branch_reset_state.set_next(cpp2::move(max_group));
+    }
+
+#line 1836 "reflect.h2"
+    auto parse_context::next_alternative() & -> void
+    {
+        cur_group_state.next_alternative();
+        cur_branch_reset_state.next_alternative();
+    }
+
+#line 1844 "reflect.h2"
+    auto parse_context::add_token(cpp2::impl::in<token_ptr> token) & -> void{
+        cur_group_state.add(token);
+    }
+
+#line 1848 "reflect.h2"
+    [[nodiscard]] auto parse_context::has_token() const& -> bool{
+        return !(cur_group_state.empty()); 
+    }
+
+#line 1852 "reflect.h2"
+    [[nodiscard]] auto parse_context::pop_token() & -> token_ptr
+    {
+        token_ptr r {nullptr}; 
+        if (has_token()) {
+            r = cur_group_state.cur_match_list.back();
+            cur_group_state.cur_match_list.pop_back();
+        }
+
+        return r; 
+    }
+
+#line 1863 "reflect.h2"
+    [[nodiscard]] auto parse_context::get_as_token() & -> token_ptr{
+        return root; 
+    }
+
+#line 1869 "reflect.h2"
+    [[nodiscard]] auto parse_context::get_cur_group() const& -> int{
+        return cur_branch_reset_state.cur_group; 
+    }
+
+#line 1873 "reflect.h2"
+    [[nodiscard]] auto parse_context::next_group() & -> int{
+        return cur_branch_reset_state.next(); 
+    }
+
+#line 1877 "reflect.h2"
+    auto parse_context::set_named_group(cpp2::impl::in<std::string> name, cpp2::impl::in<int> id) & -> void
+    {
+        if (!(named_groups.contains(name))) {// Redefinition of group name is not an error. The left most one is retained.
+            CPP2_ASSERT_IN_BOUNDS(named_groups, name) = id;
+        }
+    }
+
+#line 1884 "reflect.h2"
+    [[nodiscard]] auto parse_context::get_named_group(cpp2::impl::in<std::string> name) const& -> int
+    {
+        auto iter {named_groups.find(name)}; 
+        if (iter == named_groups.end()) {
+             return -1; 
+        }
+        else {
+            return (*cpp2::impl::assert_not_null(cpp2::move(iter))).second; 
+        }
+    }
+
+#line 1897 "reflect.h2"
+    [[nodiscard]] auto parse_context::current() const& -> char{return CPP2_ASSERT_IN_BOUNDS(regex, pos); }
+
+#line 1900 "reflect.h2"
+    [[nodiscard]] auto parse_context::get_next_position(cpp2::impl::in<bool> in_class, cpp2::impl::in<bool> no_skip) const& -> size_t
+    {
+        auto perl_syntax {false}; 
+        if (!(no_skip)) {
+            if (in_class) {
+                perl_syntax = get_modifiers().has(expression_flags::perl_code_syntax) && get_modifiers().has(expression_flags::perl_code_syntax_in_classes);
+            }
+            else {
+                perl_syntax = get_modifiers().has(expression_flags::perl_code_syntax);
+            }
+        }
+        auto cur {pos + 1}; 
+        if (cpp2::move(perl_syntax)) {
+            for( ; cpp2::impl::cmp_less(cur,regex.size()); (cur += 1) ) {
+                auto n {CPP2_ASSERT_IN_BOUNDS(regex, cur)}; 
+
+                if (space_class<char>::includes(n)) {
+                    continue;
+                }
+                else {if (!(in_class) && '#' == cpp2::move(n)) {
+                    cur = regex.find("\n", cur);
+                    if (std::string::npos == cur) {
+                        // No new line, comment runs until the end of the pattern
+                        cur = regex.size();
+                    }
+                }
+                else { // None space none comment char
+                    break;
+                }}
+            }
+        }
+
+        //  Check for end of file.
+        if (cpp2::impl::cmp_greater(cur,regex.size())) {
+            cur = regex.size();
+        }
+        return cur; 
+    }
+
+#line 1940 "reflect.h2"
+    [[nodiscard]] auto parse_context::next_impl(cpp2::impl::in<bool> in_class, cpp2::impl::in<bool> no_skip) & -> bool
+    {
+        pos = get_next_position(in_class, no_skip);
+        if (pos != regex.size()) {
+            return true; 
+        }
+        else {
+            return false; 
+        }
+    }
+
+#line 1951 "reflect.h2"
+    [[nodiscard]] auto parse_context::next() & -> auto { return next_impl(false, false); }
+#line 1952 "reflect.h2"
+    [[nodiscard]] auto parse_context::next_in_class() & -> auto { return next_impl(true, false);  }
+#line 1953 "reflect.h2"
+    [[nodiscard]] auto parse_context::next_no_skip() & -> auto { return next_impl(false, true); }
+
+#line 1955 "reflect.h2"
+    [[nodiscard]] auto parse_context::next_n(cpp2::impl::in<int> n) & -> bool{
+        auto r {true}; 
+        auto cur {0}; 
+        for( ; r && cpp2::impl::cmp_less(cur,n); (r = next()) ) {
+            cur += 1;
+        }
+        return r; 
+    }
+
+#line 1964 "reflect.h2"
+    [[nodiscard]] auto parse_context::has_next() const& -> bool{return cpp2::impl::cmp_less(pos,regex.size()); }
+
+#line 1966 "reflect.h2"
+    [[nodiscard]] auto parse_context::grab_until_impl(cpp2::impl::in<std::string> e, cpp2::impl::out<std::string> r, cpp2::impl::in<bool> any) & -> bool
+    {
+        auto end {pos}; 
+        if (any) {
+            end = regex.find_first_of(e, pos);
+        }
+        else {
+            end = regex.find(e, pos);
+        }
+
+        if (end != std::string_view::npos) {
+            r.construct(regex.substr(pos, end - pos));
+            pos = cpp2::move(end);
+            return true; 
+        }
+        else {
+            r.construct("");
+            return false; 
+        }
+    }
+
+#line 1987 "reflect.h2"
+    [[nodiscard]] auto parse_context::grab_until(cpp2::impl::in<std::string> e, cpp2::impl::out<std::string> r) & -> auto { return grab_until_impl(e, cpp2::impl::out(&r), false); }
+#line 1988 "reflect.h2"
+    [[nodiscard]] auto parse_context::grab_until(cpp2::impl::in<char> e, cpp2::impl::out<std::string> r) & -> auto { return grab_until_impl(std::string(1, e), cpp2::impl::out(&r), false); }
+#line 1989 "reflect.h2"
+    [[nodiscard]] auto parse_context::grab_until_one_of(cpp2::impl::in<std::string> e, cpp2::impl::out<std::string> r) & -> auto { return grab_until_impl(e, cpp2::impl::out(&r), true);  }
+
+#line 1991 "reflect.h2"
+    [[nodiscard]] auto parse_context::grab_n(cpp2::impl::in<int> n, cpp2::impl::out<std::string> r) & -> bool
+    {
+        if (cpp2::impl::cmp_less_eq(pos + cpp2::impl::as_<size_t>(n),regex.size())) {
+            r.construct(regex.substr(pos, cpp2::impl::as_<size_t>(n)));
+            pos += (cpp2::impl::as_<size_t>(n)) - 1;
+            return true; 
+        }
+        else {
+            r.construct("");
+            return false; 
+        }
+    }
+
+#line 2004 "reflect.h2"
+    [[nodiscard]] auto parse_context::grab_number() & -> std::string
+    {
+        auto start {pos}; 
+        auto start_search {pos}; 
+        if (CPP2_ASSERT_IN_BOUNDS(regex, start_search) == '-') {
+            start_search += 1;
+        }
+        auto end {regex.find_first_not_of("1234567890", cpp2::move(start_search))}; 
+
+        cpp2::impl::deferred_init<std::string> r; 
+        if (end != std::string::npos) {
+            r.construct(regex.substr(start, end - start));
+            pos = cpp2::move(end) - 1;
+        }
+        else {
+            r.construct(regex.substr(cpp2::move(start)));
+            pos = regex.size() - 1;
+        }
+        return cpp2::move(r.value()); 
+    }
+
+#line 2025 "reflect.h2"
+    [[nodiscard]] auto parse_context::peek_impl(cpp2::impl::in<bool> in_class) const& -> char{
+        auto next_pos {get_next_position(in_class, false)}; 
+        if (cpp2::impl::cmp_less(next_pos,regex.size())) {
+            return CPP2_ASSERT_IN_BOUNDS(regex, cpp2::move(next_pos)); 
+        }
+        else {
+            return '\0'; 
+        }
+    }
+
+#line 2035 "reflect.h2"
+    [[nodiscard]] auto parse_context::peek() const& -> auto { return peek_impl(false); }
+#line 2036 "reflect.h2"
+    [[nodiscard]] auto parse_context::peek_in_class() const& -> auto { return peek_impl(true);  }
+
+#line 2041 "reflect.h2"
+    [[nodiscard]] auto parse_context::parser_group_modifiers(cpp2::impl::in<std::string> change_str, expression_flags& parser_modifiers) & -> bool
+    {
+        auto is_negative {false}; 
+        auto is_reset {false}; 
+
+        auto apply {[&, _1 = (&is_negative), _2 = (&parser_modifiers)](cpp2::impl::in<expression_flags> flag) mutable -> void{
+            if (*cpp2::impl::assert_not_null(_1)) {
+                (*cpp2::impl::assert_not_null(_2)).clear(flag);
+            }
+            else {
+                (*cpp2::impl::assert_not_null(_2)).set(flag);
+            }
+        }}; 
+
+        auto iter {change_str.begin()}; 
+        for( ; iter != change_str.end(); (++iter) ) 
+        {
+            auto cur {*cpp2::impl::assert_not_null(iter)}; 
+            if (cur == '^') {
+                is_reset = true;
+                parser_modifiers = expression_flags::none;
+            }
+            else {if (cur == '-') {
+                if (is_reset) {static_cast<void>(error("No negative modifier allowed.")); return false; }
+                is_negative = true;
+            }
+            else {if (cur == 'i') {apply(expression_flags::case_insensitive); }
+            else {if (cur == 'm') {apply(expression_flags::multiple_lines); }
+            else {if (cur == 's') {apply(expression_flags::single_line); }
+            else {if (cur == 'n') {apply(expression_flags::no_group_captures); }
+            else {if (cur == 'x') {
+                if ((iter + 1) == change_str.end() || *cpp2::impl::assert_not_null((iter + 1)) != 'x') {
+                    // x modifier
+                    apply(expression_flags::perl_code_syntax);
+
+                    // Just x unsets xx and remove x also removes xx
+                    parser_modifiers.clear(expression_flags::perl_code_syntax_in_classes);
+                }
+                else { // xx modifier
+                    // xx also sets or unsets x
+                    apply(expression_flags::perl_code_syntax);
+                    apply(expression_flags::perl_code_syntax_in_classes);
+
+                    ++iter; // Skip the second x
+                }
+            }
+            else {
+                static_cast<void>(error("Unknown modifier: " + cpp2::to_string(cpp2::move(cur)) + "")); return false; 
+            }}}}}}}
+        }
+
+        return true; 
+    }
+
+#line 2095 "reflect.h2"
+    [[nodiscard]] auto parse_context::parse_until(cpp2::impl::in<char> term) & -> bool{
+        token_ptr cur_token {}; 
+
+        for( ; valid(); static_cast<void>(next()) ) 
+        {
+            if (term == current()) {break; }
+
+            cur_token = nullptr;
+
+            if (!(cur_token) && valid()) {cur_token = alternative_token::parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = any_token::parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = class_token::parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = escape_token_parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = global_group_reset_token_parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = group_ref_token::parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = group_token::parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = hexadecimal_token_parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = line_end_token_parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = line_start_token_parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = named_class_token_parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = octal_token_parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = range_token::parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = special_range_token::parse((*this)); }
+            if (!(cur_token) && valid()) {cur_token = word_boundary_token_parse((*this)); }
+
+            // Everything else is matched as it is.
+            if (!(cur_token) && valid()) {cur_token = char_token::parse((*this)); }
+
+            if (cur_token && valid()) {
+                add_token(cur_token);
+            }else {
+                return false; 
+            }
+        }
+
+        return true; 
+    }
+
+#line 2133 "reflect.h2"
+    [[nodiscard]] auto parse_context::parse(cpp2::impl::in<std::string> modifiers) & -> bool
+    {
+
+        expression_flags flags {}; 
+        if (!(parser_group_modifiers(modifiers, flags))) {return false; }
+        set_modifiers(cpp2::move(flags));
+
+        auto r {parse_until('\0')}; 
+        if (r) {
+            root = cur_group_state.get_as_token();
+        }
+
+        return r; 
+    }
+
+#line 2150 "reflect.h2"
+    [[nodiscard]] auto parse_context::get_pos() const& -> auto { return pos; }
+#line 2151 "reflect.h2"
+    [[nodiscard]] auto parse_context::get_range(cpp2::impl::in<size_t> start, cpp2::impl::in<size_t> end) const& -> auto { return std::string(regex.substr(start, end - start + 1));  }
+#line 2152 "reflect.h2"
+    [[nodiscard]] auto parse_context::valid() const& -> bool{return has_next() && !(has_error); }
+
+#line 2154 "reflect.h2"
+    [[nodiscard]] auto parse_context::error(cpp2::impl::in<std::string> err) & -> token_ptr{
+        has_error = true;
+        error_out("Error during parsing of regex '" + cpp2::to_string(regex) + "' at position '" + cpp2::to_string(pos) + "': " + cpp2::to_string(err) + "");
+        return nullptr; 
+    }
+
+#line 2169 "reflect.h2"
+    auto generation_function_context::add_tabs(cpp2::impl::in<int> c) & -> void{
+        int i {0}; 
+        for( ; cpp2::impl::cmp_less(i,c); i += 1 ) {
+            tabs += "  ";
+        }
+    }
+
+#line 2176 "reflect.h2"
+    auto generation_function_context::remove_tabs(cpp2::impl::in<int> c) & -> void{
+        tabs = tabs.substr(0, (cpp2::impl::as_<size_t>(c)) * 2);
+    }
+
+    generation_function_context::generation_function_context(auto const& code_, auto const& tabs_)
+                                                       : code{ code_ }
+                                                       , tabs{ tabs_ }{}
+generation_function_context::generation_function_context(){}
+
+#line 2194 "reflect.h2"
+    [[nodiscard]] auto generation_context::match_parameters() const& -> std::string{return "r.pos, ctx"; }
+
+#line 2199 "reflect.h2"
+    auto generation_context::add(cpp2::impl::in<std::string> s) & -> void{
+        auto cur {get_current()}; 
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + cpp2::to_string(s) + "\n";
+    }
+
+#line 2205 "reflect.h2"
+    auto generation_context::add_check(cpp2::impl::in<std::string> check) & -> void{
+        auto cur {get_current()}; 
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "if !cpp2::regex::" + cpp2::to_string(check) + " { r.matched = false; break; }\n";
+    }
+
+#line 2211 "reflect.h2"
+    auto generation_context::add_statefull(cpp2::impl::in<std::string> next_func, cpp2::impl::in<std::string> check) & -> void
+    {
+        end_func_statefull(check);
+
+        auto name {next_func.substr(0, next_func.size() - 2)}; 
+        start_func_named(cpp2::move(name));
+    }
+
+#line 2219 "reflect.h2"
+    auto generation_context::start_func_named(cpp2::impl::in<std::string> name) & -> void
+    {
+        auto cur {new_context()}; 
+
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + cpp2::to_string(name) + ": @struct type = {\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "  operator(): (this, cur: Iter, inout ctx: context, other) -> cpp2::regex::match_return<Iter> = {\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "    r := ctx..pass(cur);\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "    do {\n";
+        (*cpp2::impl::assert_not_null(cpp2::move(cur))).add_tabs(3);
+    }
+
+#line 2230 "reflect.h2"
+    [[nodiscard]] auto generation_context::start_func() & -> std::string
+    {
+        auto name {gen_func_name()}; 
+        start_func_named(name);
+        return cpp2::move(name) + "()"; 
+    }
+
+#line 2237 "reflect.h2"
+    auto generation_context::end_func_statefull(cpp2::impl::in<std::string> s) & -> void
+    {
+        auto cur {get_current()}; 
+        (*cpp2::impl::assert_not_null(cur)).remove_tabs(3);
+        (*cpp2::impl::assert_not_null(cur)).code += "\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "    } while false;\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "    if r.matched {\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "      r = " + cpp2::to_string(s) + ";\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "    }\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "    else {\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "      r.pos = ctx.end;\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "    }\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "    return r;\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "  }\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "}\n";
+
+        finish_context();
+    }
+
+#line 2257 "reflect.h2"
+    [[nodiscard]] auto generation_context::generate_func(cpp2::impl::in<token_ptr> token) & -> std::string
+    {
+        auto name {start_func()}; 
+        (*cpp2::impl::assert_not_null(token)).generate_code((*this));
+        end_func_statefull("other(" + cpp2::to_string(match_parameters()) + ")");
+
+        return name; 
+    }
+
+#line 2267 "reflect.h2"
+    [[nodiscard]] auto generation_context::generate_reset(cpp2::impl::in<std::set<int>> groups) & -> std::string
+    {
+        if (groups.empty()) {
+            return "cpp2::regex::no_reset()"; 
+        }
+
+        auto name {gen_reset_func_name()}; 
+        auto cur {new_context()}; 
+
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + cpp2::to_string(name) + ": @struct type = {\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "  operator(): (this, inout ctx) = {\n";
+        for ( auto const& g : groups ) {
+            (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "  ctx..set_group_invalid(" + cpp2::to_string(g) + ");\n";
+        }
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "  }\n";
+        (*cpp2::impl::assert_not_null(cur)).code += "" + cpp2::to_string((*cpp2::impl::assert_not_null(cur)).tabs) + "}\n";
+
+        finish_context();
+
+        return cpp2::move(name) + "()"; 
+    }
+
+#line 2291 "reflect.h2"
+    [[nodiscard]] auto generation_context::gen_func_name() & -> std::string{
+        auto cur_id {matcher_func}; 
+        matcher_func += 1;
+        return "func_" + cpp2::to_string(cpp2::move(cur_id)) + ""; 
+    }
+
+#line 2297 "reflect.h2"
+    [[nodiscard]] auto generation_context::next_func_name() & -> std::string{
+        return gen_func_name() + "()"; 
+    }
+
+#line 2301 "reflect.h2"
+    [[nodiscard]] auto generation_context::gen_reset_func_name() & -> std::string{
+        auto cur_id {reset_func}; 
+        reset_func += 1;
+        return "reset_" + cpp2::to_string(cpp2::move(cur_id)) + ""; 
+    }
+
+#line 2307 "reflect.h2"
+    [[nodiscard]] auto generation_context::gen_temp() & -> std::string{
+        auto cur_id {temp_name}; 
+        temp_name += 1;
+        return "tmp_" + cpp2::to_string(cpp2::move(cur_id)) + ""; 
+    }
+
+#line 2315 "reflect.h2"
+    [[nodiscard]] auto generation_context::new_context() & -> generation_function_context*{
+        gen_stack.push_back(generation_function_context());
+        auto cur {get_current()}; 
+        (*cpp2::impl::assert_not_null(cur)).tabs = "    ";
+
+        return cur; 
+    }
+
+#line 2323 "reflect.h2"
+    auto generation_context::finish_context() & -> void{
+        auto cur {get_current()}; 
+        auto base {get_base()}; 
+        (*cpp2::impl::assert_not_null(base)).code += (*cpp2::impl::assert_not_null(cpp2::move(cur))).code;
+
+        gen_stack.pop_back();
+    }
+
+#line 2333 "reflect.h2"
+    [[nodiscard]] auto generation_context::get_current() & -> generation_function_context*{
+        return &gen_stack.back(); 
+    }
+
+#line 2337 "reflect.h2"
+    [[nodiscard]] auto generation_context::get_base() & -> generation_function_context*{
+        return &CPP2_ASSERT_IN_BOUNDS_LITERAL(gen_stack, 0); 
+    }
+
+#line 2341 "reflect.h2"
+    [[nodiscard]] auto generation_context::get_entry_func() const& -> std::string{
+        return entry_func; 
+    }
+
+#line 2345 "reflect.h2"
+    [[nodiscard]] auto generation_context::create_named_group_lookup(cpp2::impl::in<std::map<std::string,int>> named_groups) const& -> std::string
+    {
+        std::string res {"get_named_group_index: (name) -> int = {\n"}; 
+
+        //  Generate if selection.
+        std::string sep {""}; 
+        for ( auto const& cur : named_groups ) {
+            res += "" + cpp2::to_string(sep) + "if name == \"" + cpp2::to_string(cur.first) + "\" { return " + cpp2::to_string(cur.second) + "; }";
+            sep = "else ";
+        }
+
+        //  Generate else branch or return if list is empty.
+        if (named_groups.empty()) {
+            res += "  _ = name;\n";
+            res += "  return -1;\n";
+        }
+        else {
+            res += "  else { return -1; }\n";
+        }
+        res += "}\n";
+        return res; 
+    }
+
+#line 2370 "reflect.h2"
+    [[nodiscard]] auto generation_context::run(cpp2::impl::in<token_ptr> token) & -> std::string{
+        entry_func = generate_func(token);
+
+        return (*cpp2::impl::assert_not_null(get_base())).code; 
+    }
+
+#line 2385 "reflect.h2"
+    alternative_token::alternative_token()
+                            : regex_token_empty{ "" }{}
+
+#line 2387 "reflect.h2"
+    [[nodiscard]] auto alternative_token::parse(parse_context& ctx) -> token_ptr{
+        if (ctx.current() != '|') {return nullptr; }
+
+        if (!(ctx.has_token())) {return ctx.error("Alternative with no content."); }
+        ctx.next_alternative();
+        return CPP2_UFCS_TEMPLATE(cpp2_new<alternative_token>)(cpp2::shared); 
+    }
+
+    alternative_token::~alternative_token() noexcept{}
+
+#line 2402 "reflect.h2"
+    alternative_token_gen::alternative_token_gen(cpp2::impl::in<token_vec> a)
+        : regex_token{ gen_string(a) }
+        , alternatives{ a }{
+
+#line 2405 "reflect.h2"
+    }
+
+#line 2407 "reflect.h2"
+    auto alternative_token_gen::generate_code(generation_context& ctx) const -> void
+    {
+        std::string functions {""}; 
+
+        for ( auto const& cur : alternatives ) {
+            std::set<int> groups {}; 
+            (*cpp2::impl::assert_not_null(cur)).add_groups(groups);
+
+            functions += ", " + ctx.generate_func(cur);
+            functions += ", " + ctx.generate_reset(cpp2::move(groups));
+        }
+
+        auto next_name {ctx.next_func_name()}; 
+
+        ctx.add_statefull(next_name, "cpp2::regex::alternative_token_matcher<char>::match(" + cpp2::to_string(ctx.match_parameters()) + ", other, " + cpp2::to_string(next_name) + " " + cpp2::to_string(cpp2::move(functions)) + ")");
+    }
+
+#line 2424 "reflect.h2"
+    auto alternative_token_gen::add_groups(std::set<int>& groups) const -> void
+    {
+        for ( auto const& cur : alternatives ) {
+            (*cpp2::impl::assert_not_null(cur)).add_groups(groups);
+        }
+    }
+
+#line 2431 "reflect.h2"
+    [[nodiscard]] auto alternative_token_gen::gen_string(cpp2::impl::in<token_vec> a) -> std::string
+    {
+        std::string r {""}; 
+        std::string sep {""}; 
+
+        for ( auto const& cur : a ) {
+            r += sep + (*cpp2::impl::assert_not_null(cur)).to_string();
+            sep = "|";
+        }
+
+        return r; 
+    }
+
+    alternative_token_gen::~alternative_token_gen() noexcept{}
+
+#line 2452 "reflect.h2"
+    any_token::any_token(cpp2::impl::in<bool> single_line)
+        : regex_token_check{ ".", "any_token_matcher<char, " + cpp2::to_string(single_line) + ">" }{
+
+#line 2454 "reflect.h2"
+    }
+
+#line 2456 "reflect.h2"
+    [[nodiscard]] auto any_token::parse(parse_context& ctx) -> token_ptr{
+        if ('.' != ctx.current()) {return nullptr; }
+
+        return CPP2_UFCS_TEMPLATE(cpp2_new<any_token>)(cpp2::shared, ctx.get_modifiers().has(expression_flags::single_line)); 
+    }
+
+    any_token::~any_token() noexcept{}
+
+#line 2473 "reflect.h2"
+    char_token::char_token(cpp2::impl::in<char> t, cpp2::impl::in<bool> ignore_case_)
+        : regex_token{ std::string(1, t) }
+        , token{ t }
+        , ignore_case{ ignore_case_ }{
+
+#line 2477 "reflect.h2"
+    }
+
+#line 2479 "reflect.h2"
+    [[nodiscard]] auto char_token::parse(parse_context& ctx) -> token_ptr{
+        return CPP2_UFCS_TEMPLATE(cpp2_new<char_token>)(cpp2::shared, ctx.current(), ctx.get_modifiers().has(expression_flags::case_insensitive)); 
+    }
+
+#line 2483 "reflect.h2"
+    auto char_token::generate_code(generation_context& ctx) const -> void
+    {
+        if (ignore_case) {
+            std::string upper {token}; 
+            std::string lower {token}; 
+{
+size_t i{0};
+
+#line 2489 "reflect.h2"
+            for( ; cpp2::impl::cmp_less(i,token.size()); i += 1 ) {
+                CPP2_ASSERT_IN_BOUNDS(lower, i) = string_util::safe_tolower(CPP2_ASSERT_IN_BOUNDS(token, i));
+                CPP2_ASSERT_IN_BOUNDS(upper, i) = string_util::safe_toupper(CPP2_ASSERT_IN_BOUNDS(token, i));
+            }
+}
+
+#line 2494 "reflect.h2"
+            if (upper != lower) {
+                gen_case_insensitive(cpp2::move(lower), cpp2::move(upper), ctx);
+            }
+            else {
+                gen_case_sensitive(ctx);
+            }
+        }
+        else {
+            gen_case_sensitive(ctx);
+        }
+    }
+
+#line 2506 "reflect.h2"
+    auto char_token::gen_case_insensitive(cpp2::impl::in<std::string> lower, cpp2::impl::in<std::string> upper, generation_context& ctx) const& -> void
+    {
+        std::string name {"str_" + cpp2::to_string(ctx.gen_temp()) + ""}; 
+        std::string lower_name {"lower_" + cpp2::to_string(name) + ""}; 
+        std::string upper_name {"upper_" + cpp2::to_string(cpp2::move(name)) + ""}; 
+        auto size {token.size()}; 
+        ctx.add("" + cpp2::to_string(lower_name) + " : std::array<char, " + cpp2::to_string(size) + " + 1> = \"" + cpp2::to_string(add_escapes(lower)) + "\";");  // TODO: Add constexpr when Issue https://github.com/hsutter/cppfront/issues/1104 is resolved.
+        ctx.add("" + cpp2::to_string(upper_name) + " : std::array<char, " + cpp2::to_string(size) + " + 1> = \"" + cpp2::to_string(add_escapes(upper)) + "\";");  // TODO: Add constexpr when Issue https://github.com/hsutter/cppfront/issues/1104 is resolved.
+        ctx.add("if std::distance(r.pos, ctx.end) < " + cpp2::to_string(size) + " {");
+        ctx.add("    r.matched = false;");
+        ctx.add("    break;");
+        ctx.add("}");
+        ctx.add("");
+        ctx.add("(copy i : int = 0) while i < " + cpp2::to_string(size) + " next (i += 1) {");
+        ctx.add("    if !(" + cpp2::to_string(cpp2::move(lower_name)) + "[i] == r.pos[i] || " + cpp2::to_string(cpp2::move(upper_name)) + "[i] == r.pos[i]) { r.matched = false; }");
+        ctx.add("}");
+        ctx.add("");
+        ctx.add("if r.matched { r.pos += " + cpp2::to_string(cpp2::move(size)) + "; }");
+        ctx.add("else { break; }");
+    }
+
+#line 2527 "reflect.h2"
+    auto char_token::gen_case_sensitive(generation_context& ctx) const& -> void
+    {
+        std::string name {"str_" + cpp2::to_string(ctx.gen_temp()) + ""}; 
+        auto size {token.size()}; 
+        ctx.add("" + cpp2::to_string(name) + " : std::array<char, " + cpp2::to_string(size) + " + 1> = \"" + cpp2::to_string(add_escapes(token)) + "\";");  // TODO: Add constexpr when Issue https://github.com/hsutter/cppfront/issues/1104 is resolved.
+        ctx.add("if std::distance(r.pos, ctx.end) < " + cpp2::to_string(size) + " {");
+        ctx.add("    r.matched = false;");
+        ctx.add("    break;");
+        ctx.add("}");
+        ctx.add("");
+        ctx.add("(copy i : int = 0) while i < " + cpp2::to_string(size) + " next (i += 1) {");
+        ctx.add("    if " + cpp2::to_string(cpp2::move(name)) + "[i] != r.pos[i] { r.matched = false; }");
+        ctx.add("}");
+        ctx.add("");
+        ctx.add("if r.matched { r.pos += " + cpp2::to_string(cpp2::move(size)) + "; }");
+        ctx.add("else { break; }");
+    }
+
+#line 2545 "reflect.h2"
+    [[nodiscard]] auto char_token::add_escapes(std::string str) const& -> std::string
+    {
+        str = string_util::replace_all(str, "\\", "\\\\");
+        str = string_util::replace_all(str, "\a", "\\a");
+        str = string_util::replace_all(str, "\f", "\\f");
+        str = string_util::replace_all(str, "\x1b", "\" \"\\x1b\" \""); // Generate a separated string. This prevents
+                                                                        // situations like `\x1bblub` from generating
+                                                                        // wrong hex characters.
+        str = string_util::replace_all(str, "\n", "\\n");
+        str = string_util::replace_all(str, "\r", "\\r");
+        str = string_util::replace_all(str, "\t", "\\t");
+
+        return cpp2::move(str); 
+    }
+
+#line 2560 "reflect.h2"
+    auto char_token::append(char_token const& that) & -> void{
+        (*this).token += that.token;
+        (*this).string_rep += that.string_rep;
+    }
+
+    char_token::~char_token() noexcept{}
+
+#line 2577 "reflect.h2"
+    class_token::class_token(cpp2::impl::in<bool> negate_, cpp2::impl::in<bool> case_insensitive_, cpp2::impl::in<std::string> class_str_, cpp2::impl::in<std::string> str)
+        : regex_token{ str }
+        , negate{ negate_ }
+        , case_insensitive{ case_insensitive_ }
+        , class_str{ class_str_ }
+#line 2578 "reflect.h2"
+    {
+
+#line 2583 "reflect.h2"
+    }
+
+#line 2586 "reflect.h2"
+    [[nodiscard]] auto class_token::parse(parse_context& ctx) -> token_ptr
+    {
+        if (ctx.current() != '[') {return nullptr; }
+
+        auto start_pos {ctx.get_pos()}; 
+
+        std::vector<std::string> supported_classes {"alnum", "alpha", "ascii", "blank", "cntrl", "digits", "graph", 
+                                                       "lower", "print", "punct", "space", "upper", "word", "xdigit"}; 
+
+        std::vector<std::string> classes {}; 
+
+        //  First step: parse until the end bracket and push single chars, ranges or groups on the class stack.
+        auto is_negate {false}; 
+        auto first {true}; 
+        auto range {false}; 
+        while( ctx.next_in_class() && (ctx.current() != ']' || first) ) 
+        {
+            if (ctx.current() == '^') 
+            {
+                is_negate = true;
+                continue; // Skip rest of the loop. Also the first update.
+            }
+
+            if (ctx.current() == '[' && ctx.peek_in_class() == ':') 
+            {
+                //  We have a character class.
+                static_cast<void>(ctx.next_n(2));// Skip [:
+
+                std::string name {""}; 
+                if (!(ctx.grab_until(":]", cpp2::impl::out(&name)))) {return ctx.error("Could not find end of character class."); }
+                if (supported_classes.end() == std::find(supported_classes.begin(), supported_classes.end(), name)) {
+                    return ctx.error("Unsupported character class. Supported ones are: " + cpp2::to_string(string_util::join(supported_classes)) + ""); 
+                }
+
+                classes.push_back("[:" + cpp2::to_string(cpp2::move(name)) + ":]");
+
+                static_cast<void>(ctx.next());// Skip ':' pointing to the ending ']'.
+            }
+            else {if (ctx.current() == '\\') 
+            {
+                if (ctx. next_no_skip()  && (ctx. current() != ']')) 
+                {
+                    if ( ' ' == ctx. current() 
+                        && ctx.get_modifiers().has(expression_flags::perl_code_syntax) 
+                        && ctx.get_modifiers().has(expression_flags::perl_code_syntax_in_classes)) 
+                    {
+                        classes.push_back(std::string(1, ctx.current()));
+                    }
+                    else {
+                        auto name {""}; 
+                        if (     'd' == ctx. current()) { name = "short_digits"; }
+                        else {if ('D' == ctx.current()) {name = "short_not_digits"; }
+                        else {if ('h' == ctx.current()) {name = "short_hor_space"; }
+                        else {if ('H' == ctx.current()) {name = "short_not_hor_space"; }
+                        else {if ('s' == ctx.current()) {name = "short_space"; }
+                        else {if ('S' == ctx.current()) {name = "short_not_space"; }
+                        else {if ('v' == ctx.current()) {name = "short_ver_space"; }
+                        else {if ('V' == ctx.current()) {name = "short_not_ver_space"; }
+                        else {if ('w' == ctx.current()) {name = "short_word"; }
+                        else {if ('W' == ctx.current()) {name = "short_not_word"; }
+                        else {
+                            return ctx.error("Unknown group escape."); 
+                        }}}}}}}}}}
+                        classes.push_back("[:" + cpp2::to_string(cpp2::move(name)) + ":]");
+                    }
+                }else {
+                    return ctx.error("Escape without a following character."); 
+                }
+            }
+            else {if (ctx.current() == '-') 
+            {
+                if (first) {// Literal if first entry.
+                    classes.push_back("" + cpp2::to_string(ctx.current()) + "");
+                }else {
+                    range = true;
+                }
+            }
+            else 
+            {
+                if (range) {// Modify last element to be a range.
+                    classes.back() += "-" + cpp2::to_string(ctx.current()) + "";
+                    range = false;
+                }
+                else {
+                    classes.push_back("" + cpp2::to_string(ctx.current()) + "");
+                }
+            }}}
+
+            first = false;
+        }
+
+        if (ctx.current() != ']') {
+            return ctx.error("Error end of character class definition before terminating ']'."); 
+        }
+        auto end_pos {ctx.get_pos()}; 
+
+        if (cpp2::move(range)) {// If '-' is last entry treat it as a literal char.
+            classes.push_back("-");
+        }
+
+        // Second step: Wrap the item on the class stack with corresponding class implementation.
+        for ( auto& cur : classes ) 
+        {
+            if (cur.starts_with("[:")) {
+                auto name {cur.substr(2, cur.size() - 4)}; 
+                cur = create_matcher("" + cpp2::to_string(cpp2::move(name)) + "_class", "");
+            }
+            else {if (1 != cur.size()) {
+                cur = create_matcher("range_class_entry", "'" + cpp2::to_string(CPP2_ASSERT_IN_BOUNDS_LITERAL(cur, 0)) + "', '" + cpp2::to_string(CPP2_ASSERT_IN_BOUNDS_LITERAL(cur, 2)) + "'");
+            }
+            else {
+                cur = create_matcher("single_class_entry", "'" + cpp2::to_string(cur) + "'");
+            }}
+        }
+
+        auto inner {string_util::join(cpp2::move(classes))}; 
+        auto string_rep {ctx.get_range(cpp2::move(start_pos), cpp2::move(end_pos))}; 
+        return CPP2_UFCS_TEMPLATE(cpp2_new<class_token>)(cpp2::shared, 
+            cpp2::move(is_negate), 
+            ctx.get_modifiers().has(expression_flags::case_insensitive), 
+            cpp2::move(inner), 
+            cpp2::move(string_rep)
+            ); 
+    }
+
+#line 2711 "reflect.h2"
+    auto class_token::generate_code(generation_context& ctx) const -> void
+    {
+        ctx.add_check("class_token_matcher<char, " + cpp2::to_string(negate) + ", " + cpp2::to_string(case_insensitive) + ", " + cpp2::to_string(class_str) + ">::match(" + cpp2::to_string(ctx.match_parameters()) + ")");
+    }
+
+#line 2716 "reflect.h2"
+    [[nodiscard]] auto class_token::create_matcher(cpp2::impl::in<std::string> name, cpp2::impl::in<std::string> template_arguments) -> std::string
+    {
+        auto sep {", "}; 
+        if (template_arguments.empty()) {sep = ""; }
+
+        return "::cpp2::regex::" + cpp2::to_string(name) + "<char" + cpp2::to_string(cpp2::move(sep)) + cpp2::to_string(template_arguments) + ">"; 
+    }
+
+    class_token::~class_token() noexcept{}
+
+#line 2728 "reflect.h2"
+[[nodiscard]] auto escape_token_parse(parse_context& ctx) -> token_ptr
+{
+    if (ctx.current() != '\\') {return nullptr; }
+
+#line 2733 "reflect.h2"
+    if (std::string::npos == std::string("afenrt^.[]()*{}?+|\\").find(ctx.peek())) {
+        return nullptr; 
+    }
+
+    static_cast<void>(ctx.next());// Skip escape
+
+    if (std::string::npos != std::string("afenrt\\").find(ctx.current())) 
+    {
+        // Escape of string special char
+        char t {'\0'}; 
+        if (     'a' == ctx. current()) { t = '\a'; }
+        else {if ('f' == ctx.current()) {t = '\f'; }
+        else {if ('e' == ctx.current()) {t = '\x1b'; }
+        else {if ('n' == ctx.current()) {t = '\n'; }
+        else {if ('r' == ctx.current()) {t = '\r'; }
+        else {if ('t' == ctx.current()) {t = '\t'; }
+        else {if ('\\' == ctx.current()) {t = '\\'; }
+        else {return ctx.error("Internal: missing switch case for special escape."); }}}}}}}
+
+        auto r {CPP2_UFCS_TEMPLATE(cpp2_new<char_token>)(cpp2::shared, cpp2::move(t), false)}; 
+        (*cpp2::impl::assert_not_null(r)).set_string("\\" + cpp2::to_string(ctx.current()) + "");
+        return r; 
+    }
+    else 
+    {
+        // Escape of regex special char
+        auto r {CPP2_UFCS_TEMPLATE(cpp2_new<char_token>)(cpp2::shared, ctx.current(), false)}; 
+        (*cpp2::impl::assert_not_null(r)).set_string("\\" + cpp2::to_string(ctx.current()) + "");
+        return r; 
+    }
+
+}
+
+#line 2769 "reflect.h2"
+[[nodiscard]] auto global_group_reset_token_parse(parse_context& ctx) -> token_ptr
+{
+    if (!((ctx.current() == '\\' && ctx.peek() == 'K'))) {return nullptr; }
+
+    static_cast<void>(ctx.next());// Skip escape.
+    return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_code>)(cpp2::shared, "\\K", "ctx..set_group_start(0, r.pos);"); 
+}
+
+#line 2791 "reflect.h2"
+    group_ref_token::group_ref_token(cpp2::impl::in<int> id_, cpp2::impl::in<bool> case_insensitive_, cpp2::impl::in<std::string> str)
+        : regex_token{ str }
+        , id{ id_ }
+        , case_insensitive{ case_insensitive_ }
+#line 2792 "reflect.h2"
+    {
+
+#line 2796 "reflect.h2"
+    }
+
+#line 2798 "reflect.h2"
+    [[nodiscard]] auto group_ref_token::parse(parse_context& ctx) -> token_ptr
+    {
+        if (ctx.current() != '\\') {return nullptr; }
+
+        std::string str {"\\"}; 
+        std::string group {""}; 
+
+        if ([_0 = '0', _1 = ctx.peek(), _2 = '9']{ return cpp2::impl::cmp_less_eq(_0,_1) && cpp2::impl::cmp_less_eq(_1,_2); }()) 
+        {
+            static_cast<void>(ctx.next());// Skip escape
+            group = ctx.grab_number();
+
+            if (cpp2::impl::cmp_greater_eq(group.size(),cpp2::impl::as_<size_t, 3>())) 
+            {
+                // Octal syntax (\000) not a group ref matcher.
+                auto number {0}; 
+                if (!(string_util::string_to_int(group, number, 8))) {return ctx.error("Could not convert octal to int."); }
+
+                char number_as_char {cpp2::unsafe_narrow<char>(cpp2::move(number))}; 
+
+                auto token {CPP2_UFCS_TEMPLATE(cpp2_new<char_token>)(cpp2::shared, number_as_char, ctx.get_modifiers().has(expression_flags::case_insensitive))}; 
+                (*cpp2::impl::assert_not_null(token)).set_string("\\" + cpp2::to_string(string_util::int_to_string<8>(cpp2::impl::as_<int>(cpp2::move(number_as_char)))) + "");
+
+                return token; 
+            }
+
+            str += group;
+            // Regular group ref
+        }
+        else {if ('g' == ctx.peek()) 
+        {
+            static_cast<void>(ctx.next());// Skip escape
+            if (!(ctx.next())) {return ctx.error("Group escape without a following char."); }// Skip g
+
+            str += "g";
+
+            if (ctx.current() == '{') {
+                str += "{";
+                if (!((ctx.next() && ctx.grab_until('}', cpp2::impl::out(&group))))) {return ctx.error("No ending bracket."); }
+
+                str += group + "}";
+            }
+            else {
+                group = ctx.grab_number();
+                str += group;
+            }
+        }
+        else {if ('k' == ctx.peek()) 
+        {
+            static_cast<void>(ctx.next());// Skip escape
+            if (!(ctx.next())) {return ctx.error("Group escape without a following char."); }// Skip k
+
+            str += "k";
+
+            auto term_char {'\0'}; 
+            if (ctx.current() == '{') {term_char = '}'; }
+            else {if (ctx.current() == '<') {term_char = '>'; }
+            else {if (ctx.current() == '\'') {term_char = '\''; }
+            else {
+                return ctx.error("Group escape has wrong operator."); 
+            }}}
+
+            str += ctx.current();
+
+            if (!((ctx.next() && ctx.grab_until(term_char, cpp2::impl::out(&group))))) {return ctx.error("No ending bracket."); }
+
+            str += group + cpp2::move(term_char);
+        }
+        else 
+        {
+            // No group ref matcher
+            return nullptr; 
+        }}}
+
+        // Parse the group
+        group = string_util::trim_copy(group);
+        int group_id {0}; 
+        if (string_util::string_to_int(group, group_id)) 
+        {
+            if (cpp2::impl::cmp_less(group_id,0)) {
+                group_id = ctx.get_cur_group() + group_id;
+
+                if (cpp2::impl::cmp_less(group_id,1)) {// Negative and zero are no valid groups.
+                    return ctx.error("Relative group reference does not reference a valid group. (Would be " + cpp2::to_string(group_id) + ".)"); 
+                }
+            }
+
+            if (cpp2::impl::cmp_greater_eq(group_id,ctx.get_cur_group())) {
+                return ctx.error("Group reference is used before the group is declared."); 
+            }
+        }
+        else 
+        {
+            // Named group
+            group_id = ctx.get_named_group(group);
+            if (-1 == group_id) {return ctx.error("Group names does not exist. (Name is: " + cpp2::to_string(cpp2::move(group)) + ")"); }
+        }
+
+        return CPP2_UFCS_TEMPLATE(cpp2_new<group_ref_token>)(cpp2::shared, cpp2::move(group_id), ctx.get_modifiers().has(expression_flags::case_insensitive), cpp2::move(str)); 
+    }
+
+#line 2899 "reflect.h2"
+    auto group_ref_token::generate_code(generation_context& ctx) const -> void{
+        ctx.add_check("group_ref_token_matcher<char, " + cpp2::to_string(id) + ", " + cpp2::to_string(case_insensitive) + ">(" + cpp2::to_string(ctx.match_parameters()) + ")");
+    }
+
+    group_ref_token::~group_ref_token() noexcept{}
+
+#line 2922 "reflect.h2"
+    [[nodiscard]] auto group_token::parse_lookahead(parse_context& ctx, cpp2::impl::in<std::string> syntax, cpp2::impl::in<bool> positive) -> token_ptr
+    {
+        static_cast<void>(ctx.next());// Skip last token defining the syntax
+
+        auto r {CPP2_UFCS_TEMPLATE(cpp2_new<lookahead_token>)(cpp2::shared, positive)}; 
+
+        auto old_state {ctx.start_group()}; 
+        if (!(ctx.parse_until(')'))) {return ctx.error("Lookahead without a closing bracket."); }
+        (*cpp2::impl::assert_not_null(r)).inner = ctx.end_group(cpp2::move(old_state));
+        (*cpp2::impl::assert_not_null(r)).set_string("(" + cpp2::to_string(syntax) + cpp2::to_string((*cpp2::impl::assert_not_null((*cpp2::impl::assert_not_null(r)).inner)).to_string()) + ")");
+
+        return r; 
+    }
+
+#line 2936 "reflect.h2"
+    [[nodiscard]] auto group_token::parse(parse_context& ctx) -> token_ptr
+    {
+        if (ctx.current() != '(') {return nullptr; }
+
+        auto has_id {!(ctx.get_modifiers().has(expression_flags::no_group_captures))}; 
+        auto has_pattern {true}; 
+        std::string group_name {""}; 
+        auto group_name_brackets {true}; 
+        std::string modifiers {""}; 
+        auto modifiers_change_to {ctx.get_modifiers()}; 
+
+        //  Skip the '('
+        if (!(ctx.next())) {return ctx.error("Group without closing bracket."); }
+
+        if (ctx.current() == '?') 
+        {
+            //  Special group
+            if (!(ctx.next_no_skip())) {return ctx.error("Missing character after group opening."); }
+
+            if (ctx.current() == '<' || ctx.current() == '\'') 
+            {
+                //  Named group
+                auto end_char {ctx.current()}; 
+                if (end_char == '<') {
+                    end_char = '>';
+                }else {
+                    group_name_brackets = false;
+                }
+                has_id = true; // Force id for named groups.
+                if (!(ctx.next()))            { return ctx. error("Missing ending bracket for named group.");  }/* skip '<' */
+                if (!(ctx.grab_until(cpp2::move(end_char), cpp2::impl::out(&group_name)))) {return ctx.error("Missing ending bracket for named group."); }
+                if (!(ctx.next())) {return ctx.error("Group without closing bracket."); }
+            }
+            else {if (ctx.current() == '#') 
+            {
+                //  Comment
+                std::string comment_str {""}; 
+                static_cast<void>(ctx.next());// Skip #
+                if (!(ctx.grab_until(")", cpp2::impl::out(&comment_str)))) {return ctx.error("Group without closing bracket."); }
+                //  Do not add comment. Has problems with ranges.
+
+                //  Pop token and add a list. This fixes comments between a token and a range
+                if (ctx.has_token()) {
+                    token_vec list {}; 
+                    list.push_back(ctx.pop_token());
+                    list.push_back(CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_empty>)(cpp2::shared, "(?#" + cpp2::to_string(cpp2::move(comment_str)) + ")"));
+
+                    return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_list>)(cpp2::shared, cpp2::move(list)); 
+                }
+                else {
+                    return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_empty>)(cpp2::shared, "(?#" + cpp2::to_string(cpp2::move(comment_str)) + ")"); 
+                }
+            }
+            else {if (ctx.current() == '|') 
+            {
+                //  Branch reset group
+
+                if (!(ctx.next()))            { return ctx. error("Missing ending bracket for named group.");  }/* skip '|' */
+
+                auto old_parser_state {ctx.start_group()}; 
+                auto old_branch_state {ctx.branch_reset_new_state()}; 
+                if (!(ctx.parse_until(')'))) {return nullptr; }
+                ctx.branch_reset_restore_state(cpp2::move(old_branch_state));
+                auto inner_ {ctx.end_group(cpp2::move(old_parser_state))}; 
+
+                token_vec list {CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_empty>)(cpp2::shared, "(?|"), cpp2::move(inner_), CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_empty>)(cpp2::shared, ")")}; 
+                return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_list>)(cpp2::shared, cpp2::move(list)); 
+            }
+            else {if (ctx.current() == '=' || ctx.current() == '!') 
+            {
+                return parse_lookahead(ctx, "?" + cpp2::to_string(ctx.current()) + "", ctx.current() == '='); 
+            }
+            else 
+            {
+                //  Simple modifier
+                has_id = false;
+                if (!(ctx.grab_until_one_of("):", cpp2::impl::out(&modifiers)))) {return ctx.error("Missing ending bracket for group."); }
+                if (!(ctx.parser_group_modifiers(modifiers, modifiers_change_to))) {
+                     return nullptr; 
+                }
+
+                if (')' == ctx.current()) {
+                    has_pattern = false;
+                }
+                else {
+                    if (!(ctx.next()))            { return ctx. error("Missing ending bracket for group.");  }/* skip ':' */
+                }
+            }}}}
+        }
+        else {if (ctx.current() == '*') 
+        {
+            //  Named pattern
+            static_cast<void>(ctx.next());// Skip *.
+            std::string name {""}; 
+            if (!(ctx.grab_until(':', cpp2::impl::out(&name)))) {return ctx.error("Missing colon for named pattern."); }
+
+            if (name == "pla" || name == "positive_lookahead") {
+                return parse_lookahead(ctx, "*" + cpp2::to_string(cpp2::move(name)) + ":", true); 
+            }
+            else {if (name == "nla" || name == "negative_lookahead") {
+                return parse_lookahead(ctx, "*" + cpp2::to_string(cpp2::move(name)) + ":", false); 
+            }
+            else {
+                return ctx.error("Unknown named group pattern: '" + cpp2::to_string(cpp2::move(name)) + "'"); 
+            }}
+        }}
+
+        if (cpp2::move(has_pattern)) 
+        {
+            //  Regular group
+
+            auto r {CPP2_UFCS_TEMPLATE(cpp2_new<group_token>)(cpp2::shared)}; 
+            if (has_id) {
+                (*cpp2::impl::assert_not_null(r)).number = ctx.next_group();
+
+                if (0 != group_name.size()) {
+                    ctx.set_named_group(group_name, (*cpp2::impl::assert_not_null(r)).number);
+                }
+            }
+
+            auto old_state {ctx.start_group()}; 
+            ctx.set_modifiers(cpp2::move(modifiers_change_to));
+            if (!(ctx.parse_until(')'))) {return nullptr; }
+            (*cpp2::impl::assert_not_null(r)).inner = ctx.end_group(cpp2::move(old_state));
+            (*cpp2::impl::assert_not_null(r)).set_string(gen_string(cpp2::move(group_name), cpp2::move(group_name_brackets), !(cpp2::move(has_id)), cpp2::move(modifiers), (*cpp2::impl::assert_not_null(r)).inner));
+
+            return r; 
+        }
+        else 
+        {
+            //  Only a modifier
+            ctx.set_modifiers(cpp2::move(modifiers_change_to));
+
+            return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_empty>)(cpp2::shared, "(?" + cpp2::to_string(cpp2::move(modifiers)) + ")"); 
+        }
+    }
+
+#line 3073 "reflect.h2"
+    [[nodiscard]] auto group_token::gen_string(cpp2::impl::in<std::string> name, cpp2::impl::in<bool> name_brackets, cpp2::impl::in<bool> has_modifier, cpp2::impl::in<std::string> modifiers, cpp2::impl::in<token_ptr> inner_) -> std::string
+    {
+        std::string start {"("}; 
+        if (0 != name.size()) {
+            if (name_brackets) {
+                start += "?<" + cpp2::to_string(name.data()) + ">";
+            }
+            else {
+                start += "?'" + cpp2::to_string(name.data()) + "'";
+            }
+        }
+        else {if (has_modifier) {
+            start += "?" + modifiers + ":";
+        }}
+
+        return cpp2::move(start) + (*cpp2::impl::assert_not_null(inner_)).to_string() + ")"; 
+    }
+
+#line 3091 "reflect.h2"
+    auto group_token::generate_code(generation_context& ctx) const -> void
+    {
+        if (-1 != number) {
+            ctx.add("ctx..set_group_start(" + cpp2::to_string(number) + ", r.pos);");
+        }
+
+        (*cpp2::impl::assert_not_null(inner)).generate_code(ctx);
+        if (-1 != number) {
+            ctx.add("ctx..set_group_end(" + cpp2::to_string(number) + ", r.pos);");
+            auto tmp_name {ctx.gen_temp()}; 
+            ctx.add("" + cpp2::to_string(tmp_name) + "_func :=  :() = {");
+            ctx.add("  if !r&$*.matched {");
+            ctx.add("    ctx&$*..set_group_invalid(" + cpp2::to_string(number) + ");");
+            ctx.add("  }");
+            ctx.add("};");
+            ctx.add("" + cpp2::to_string(tmp_name) + " :=  cpp2::regex::make_on_return(" + cpp2::to_string(tmp_name) + "_func);");
+            ctx.add("_ = " + cpp2::to_string(cpp2::move(tmp_name)) + ";");// Logic is done in the destructor. Same behavior as for guard objects.
+        }
+    }
+
+#line 3111 "reflect.h2"
+    auto group_token::add_groups(std::set<int>& groups) const -> void
+    {
+        (*cpp2::impl::assert_not_null(inner)).add_groups(groups);
+        if (-1 != number) {
+            static_cast<void>(groups.insert(number));
+        }
+    }
+
+    group_token::~group_token() noexcept{}
+
+#line 3123 "reflect.h2"
+[[nodiscard]] auto hexadecimal_token_parse(parse_context& ctx) -> token_ptr
+{
+    if (!((ctx.current() == '\\' && ctx.peek() == 'x'))) {return nullptr; }
+
+    static_cast<void>(ctx.next());// Skip escape.
+
+    if (!(ctx.next())) {return ctx.error("x escape without number."); }
+
+    auto has_brackets {false}; 
+    std::string number_str {""}; 
+    if ('{' == ctx.current()) {
+        //  Bracketed
+        has_brackets = true;
+        static_cast<void>(ctx.next());// Skip '{'
+        if (!(ctx.grab_until('}', cpp2::impl::out(&number_str)))) {return ctx.error("No ending bracket for \\x"); }
+    }
+    else {
+        //  Grab two chars
+        if (!(ctx.grab_n(2, cpp2::impl::out(&number_str)))) {return ctx.error("Missing hexadecimal digits after \\x."); }
+    }
+
+    auto number {0}; 
+    if (!(string_util::string_to_int(cpp2::move(number_str), number, 16))) {return ctx.error("Could not convert hexadecimal to int."); }
+
+    // TODO: Change for unicode.
+    char number_as_char {cpp2::unsafe_narrow<char>(cpp2::move(number))}; 
+
+    std::string syntax {string_util::int_to_string<16>(cpp2::impl::as_<int>(number_as_char))}; 
+    if (cpp2::move(has_brackets)) {
+        syntax = "{" + cpp2::to_string(syntax) + "}";
+    }
+    syntax = "\\x" + cpp2::to_string(syntax) + "";
+
+    auto r {CPP2_UFCS_TEMPLATE(cpp2_new<char_token>)(cpp2::shared, cpp2::move(number_as_char), ctx.get_modifiers().has(expression_flags::case_insensitive))}; 
+    (*cpp2::impl::assert_not_null(r)).set_string(cpp2::move(syntax));
+    return r; 
+}
+
+#line 3164 "reflect.h2"
+[[nodiscard]] auto line_end_token_parse(parse_context& ctx) -> token_ptr
+{
+    if (ctx.current() == '$' || (ctx.current() == '\\' && ctx.peek() == '$')) {
+        if ((ctx.current() == '\\')) {static_cast<void>(ctx.next()); }// Skip escape
+        return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_check>)(cpp2::shared, "$", "line_end_token_matcher<char, " + cpp2::to_string(ctx.get_modifiers().has(expression_flags::multiple_lines)) + ", true>"); 
+    }
+    else {if (ctx.current() == '\\' && (ctx.peek() == 'z' || ctx.peek() == 'Z')) {
+        static_cast<void>(ctx.next());// Skip escape
+
+        auto negate {ctx.current() == 'Z'}; 
+        return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_check>)(cpp2::shared, "\\" + cpp2::to_string(ctx.current()) + "", "line_end_token_matcher<char, false, " + cpp2::to_string(cpp2::move(negate)) + ">"); 
+    }
+    else {
+        return nullptr; 
+    }}
+}
+
+#line 3184 "reflect.h2"
+[[nodiscard]] auto line_start_token_parse(parse_context& ctx) -> token_ptr
+{
+    if (ctx.current() != '^' && !((ctx.current() == '\\' && ctx.peek() == 'A'))) {return nullptr; }
+
+    if (ctx.current() == '\\') {
+        static_cast<void>(ctx.next());
+        return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_check>)(cpp2::shared, "\\A", "line_start_token_matcher<char, false>"); 
+    }
+    else {
+        return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_check>)(cpp2::shared, "^", "line_start_token_matcher<char, " + cpp2::to_string(ctx.get_modifiers().has(expression_flags::multiple_lines)) + ">"); 
+    }
+}
+
+#line 3209 "reflect.h2"
+    lookahead_token::lookahead_token(cpp2::impl::in<bool> positive_)
+        : regex_token{ "" }
+        , positive{ positive_ }{
+
+#line 3211 "reflect.h2"
+    }
+
+#line 3213 "reflect.h2"
+    auto lookahead_token::generate_code(generation_context& ctx) const -> void{
+        auto inner_name {ctx.generate_func(inner)}; 
+
+        ctx.add_check("lookahead_token_matcher<char, " + cpp2::to_string(positive) + ">(" + cpp2::to_string(ctx.match_parameters()) + ", " + cpp2::to_string(cpp2::move(inner_name)) + ")");
+    }
+
+#line 3219 "reflect.h2"
+    auto lookahead_token::add_groups(std::set<int>& groups) const -> void{
+        (*cpp2::impl::assert_not_null(inner)).add_groups(groups);
+    }
+
+    lookahead_token::~lookahead_token() noexcept{}
+
+#line 3227 "reflect.h2"
+[[nodiscard]] auto named_class_token_parse(parse_context& ctx) -> token_ptr
+{
+    if (ctx.current() != '\\') {return nullptr; }
+
+    auto name {""}; 
+    auto c_next {ctx.peek()}; 
+
+    if (     'd' == c_next) { name = "named_class_digits"; }
+    else {if ('D' == c_next) {name = "named_class_not_digits"; }
+    else {if ('h' == c_next) {name = "named_class_hor_space"; }
+    else {if ('H' == c_next) {name = "named_class_not_hor_space"; }
+    else {if ('N' == c_next) {name = "named_class_no_new_line"; }
+    else {if ('s' == c_next) {name = "named_class_space"; }
+    else {if ('S' == c_next) {name = "named_class_not_space"; }
+    else {if ('v' == c_next) {name = "named_class_ver_space"; }
+    else {if ('V' == c_next) {name = "named_class_not_ver_space"; }
+    else {if ('w' == c_next) {name = "named_class_word"; }
+    else {if ('W' == cpp2::move(c_next)) {name = "named_class_not_word"; }
+    else                  { return nullptr; }}}}}}}}}}}
+
+    static_cast<void>(ctx.next());// Skip escape
+
+    return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_check>)(cpp2::shared, "\\" + cpp2::to_string(ctx.current()) + "", "" + cpp2::to_string(cpp2::move(name)) + "<char, " + cpp2::to_string(ctx.get_modifiers().has(expression_flags::case_insensitive)) + ">::match"); 
+}
+
+#line 3255 "reflect.h2"
+[[nodiscard]] auto octal_token_parse(parse_context& ctx) -> token_ptr
+{
+    if (!((ctx.current() == '\\' && ctx.peek() == 'o'))) {return nullptr; }
+
+    static_cast<void>(ctx.next());// Skip escape.
+
+    if (!(ctx.next()))      { return ctx. error("o escape without number."); }
+    if (ctx.current() != '{') {return ctx.error("Missing opening bracket for \\o."); }
+
+    std::string number_str {""}; 
+    static_cast<void>(ctx.next());// Skip '{'
+    if (!(ctx.grab_until('}', cpp2::impl::out(&number_str)))) {return ctx.error("No ending bracket for \\o"); }
+
+    auto number {0}; 
+    if (!(string_util::string_to_int(cpp2::move(number_str), number, 8))) {return ctx.error("Could not convert octal to int."); }
+
+    // TODO: Change for unicode.
+    char number_as_char {cpp2::unsafe_narrow<char>(cpp2::move(number))}; 
+
+    std::string syntax {"\\o{" + cpp2::to_string(string_util::int_to_string<8>(cpp2::impl::as_<int>(number_as_char))) + "}"}; 
+    auto r {CPP2_UFCS_TEMPLATE(cpp2_new<char_token>)(cpp2::shared, cpp2::move(number_as_char), ctx.get_modifiers().has(expression_flags::case_insensitive))}; 
+    (*cpp2::impl::assert_not_null(r)).set_string(cpp2::move(syntax));
+    return r; 
+}
+
+#line 3292 "reflect.h2"
+    range_token::range_token()
+                             : regex_token{ "" }{}
+
+#line 3294 "reflect.h2"
+    [[nodiscard]] auto range_token::parse(parse_context& ctx) -> token_ptr
+    {
+        auto r {CPP2_UFCS_TEMPLATE(cpp2_new<range_token>)(cpp2::shared)}; 
+        if (ctx.current() == '{') 
+        {
+            if (!(ctx.has_token())) {return ctx.error("'{' without previous element."); }
+
+            std::string inner {""}; 
+            if (!(ctx.grab_until('}', cpp2::impl::out(&inner)))) {return ctx.error("Missing closing bracket '}'."); }
+
+            inner = string_util::trim_copy(inner.substr(1));  // Remove '{' and white spaces.
+            if (inner.empty()) {return ctx.error("Empty range specifier. Either '{n}', '{n,}', '{,m}' '{n,m}'"); }
+
+            // Non-greedy or possessive
+            (*cpp2::impl::assert_not_null(r)).parse_modifier(ctx);
+
+            // Get range arguments
+            std::string min_count_str {"-1"}; 
+            std::string max_count_str {"-1"}; 
+
+            size_t sep {inner.find(",")}; 
+            if (sep == std::string::npos) 
+            {
+                min_count_str = inner;
+                max_count_str = inner;
+                if (!(string_util::string_to_int(cpp2::move(inner), (*cpp2::impl::assert_not_null(r)).min_count))) {return ctx.error("Could not convert range to number."); }
+                (*cpp2::impl::assert_not_null(r)).max_count = (*cpp2::impl::assert_not_null(r)).min_count;
+            }
+            else 
+            {
+                std::string inner_first {string_util::trim_copy(inner.substr(0, sep))}; 
+                std::string inner_last {string_util::trim_copy(cpp2::move(inner).substr(cpp2::move(sep) + 1))}; 
+
+                if ((inner_first.empty() && inner_last.empty())) {
+                    return ctx.error("Empty range specifier. Either '{n}', '{n,}', '{,m}' '{n,m}'"); 
+                }
+
+                if (!(inner_first.empty())) {
+                    min_count_str = inner_first;
+                    if (!(string_util::string_to_int(cpp2::move(inner_first), (*cpp2::impl::assert_not_null(r)).min_count))) {return ctx.error("Could not convert range to number."); }
+                }
+                if (!(inner_last.empty())) {
+                    max_count_str = inner_last;
+                    if (!(string_util::string_to_int(cpp2::move(inner_last), (*cpp2::impl::assert_not_null(r)).max_count))) {return ctx.error("Could not convert range to number."); }
+                }
+            }
+
+            // Check validity of the range.
+            if (-1 != (*cpp2::impl::assert_not_null(r)).min_count) {
+                if (!((cpp2::impl::cmp_less_eq(0,(*cpp2::impl::assert_not_null(r)).min_count)))) {
+                    return ctx.error("Min value in range is negative. Have " + cpp2::to_string((*cpp2::impl::assert_not_null(r)).min_count) + ")"); 
+                }
+            }
+            if (-1 != (*cpp2::impl::assert_not_null(r)).max_count) {
+                if (!((cpp2::impl::cmp_less_eq(0,(*cpp2::impl::assert_not_null(r)).max_count)))) {
+                    return ctx.error("Max value in range is negative. Have " + cpp2::to_string((*cpp2::impl::assert_not_null(r)).max_count) + ")"); 
+                }
+                if (-1 != (*cpp2::impl::assert_not_null(r)).min_count) {
+                    if (!((cpp2::impl::cmp_less_eq((*cpp2::impl::assert_not_null(r)).min_count,(*cpp2::impl::assert_not_null(r)).max_count)))) {
+                        return ctx.error("Min and max values in range are wrong it should hold 0 <= min <= max. Have 0 <= " + cpp2::to_string((*cpp2::impl::assert_not_null(r)).min_count) + " <= " + cpp2::to_string((*cpp2::impl::assert_not_null(r)).max_count) + ""); 
+                    }
+                }
+            }
+
+            (*cpp2::impl::assert_not_null(r)).inner_token = ctx.pop_token();
+            (*cpp2::impl::assert_not_null(r)).string_rep = (*cpp2::impl::assert_not_null((*cpp2::impl::assert_not_null(r)).inner_token)).to_string() + (*cpp2::impl::assert_not_null(r)).gen_range_string() + (*cpp2::impl::assert_not_null(r)).gen_mod_string();
+
+            return r; 
+        }
+
+        return nullptr; 
+    }
+
+#line 3367 "reflect.h2"
+    auto range_token::parse_modifier(parse_context& ctx) & -> void
+    {
+        if (ctx.peek() == '?') {
+            kind = range_flags::not_greedy;
+            static_cast<void>(ctx.next());
+        }
+        else {if (ctx.peek() == '+') {
+            kind = range_flags::possessive;
+            static_cast<void>(ctx.next());
+        }}
+    }
+
+#line 3379 "reflect.h2"
+    [[nodiscard]] auto range_token::gen_mod_string() const& -> std::string
+    {
+        if (kind == range_flags::not_greedy) {
+            return "?"; 
+        }
+        else {if (kind == range_flags::possessive) {
+            return "+"; 
+        }
+        else {
+            return ""; 
+        }}
+    }
+
+#line 3392 "reflect.h2"
+    [[nodiscard]] auto range_token::gen_range_string() const& -> std::string
+    {
+        std::string r {""}; 
+        if (min_count == max_count) {
+            r += "{" + cpp2::to_string(min_count) + "}";
+        }
+        else {if (min_count == -1) {
+            r += "{," + cpp2::to_string(max_count) + "}";
+        }
+        else {if (max_count == -1) {
+            r += "{" + cpp2::to_string(min_count) + ",}";
+        }
+        else {
+            r += "{" + cpp2::to_string(min_count) + "," + cpp2::to_string(max_count) + "}";
+        }}}
+
+        return r; 
+    }
+
+#line 3411 "reflect.h2"
+    auto range_token::generate_code(generation_context& ctx) const -> void
+    {
+        auto inner_name {ctx.generate_func(inner_token)}; 
+        std::set<int> groups {}; 
+        (*cpp2::impl::assert_not_null(inner_token)).add_groups(groups);
+        auto reset_name {ctx.generate_reset(cpp2::move(groups))}; 
+
+        auto next_name {ctx.next_func_name()}; 
+        ctx.add_statefull(next_name, "cpp2::regex::range_token_matcher<char, " + cpp2::to_string(min_count) + ", " + cpp2::to_string(max_count) + ", " + cpp2::to_string(kind) + ">::match(" + cpp2::to_string(ctx.match_parameters()) + ", " + cpp2::to_string(cpp2::move(inner_name)) + ", " + cpp2::to_string(cpp2::move(reset_name)) + ", other, " + cpp2::to_string(next_name) + ")");
+    }
+
+#line 3422 "reflect.h2"
+    auto range_token::add_groups(std::set<int>& groups) const -> void{
+        (*cpp2::impl::assert_not_null(inner_token)).add_groups(groups);
+    }
+
+    range_token::~range_token() noexcept{}
+
+#line 3435 "reflect.h2"
+    [[nodiscard]] auto special_range_token::parse(parse_context& ctx) -> token_ptr
+    {
+        auto r {CPP2_UFCS_TEMPLATE(cpp2_new<special_range_token>)(cpp2::shared)}; 
+        char symbol {'\0'}; 
+        if (ctx.current() == '*') {
+            (*cpp2::impl::assert_not_null(r)).min_count = 0;
+            (*cpp2::impl::assert_not_null(r)).max_count = -1;
+            symbol = '*';
+        }
+        else {if (ctx.current() == '+') {
+            (*cpp2::impl::assert_not_null(r)).min_count = 1;
+            (*cpp2::impl::assert_not_null(r)).max_count = -1;
+            symbol = '+';
+        }else {if (ctx.current() == '?') {
+            (*cpp2::impl::assert_not_null(r)).min_count = 0;
+            (*cpp2::impl::assert_not_null(r)).max_count = 1;
+            symbol = '?';
+        }else {
+            return nullptr; 
+        }}}
+
+        if (!(ctx.has_token())) {return ctx.error("'" + cpp2::to_string(ctx.current()) + "' without previous element."); }
+
+#line 3459 "reflect.h2"
+        (*cpp2::impl::assert_not_null(r)).parse_modifier(ctx);
+
+        (*cpp2::impl::assert_not_null(r)).inner_token = ctx.pop_token();
+        (*cpp2::impl::assert_not_null(r)).string_rep = (*cpp2::impl::assert_not_null((*cpp2::impl::assert_not_null(r)).inner_token)).to_string() + cpp2::move(symbol) + (*cpp2::impl::assert_not_null(r)).gen_mod_string();
+        return r; 
+    }
+
+    special_range_token::~special_range_token() noexcept{}
+
+#line 3472 "reflect.h2"
+[[nodiscard]] auto word_boundary_token_parse(parse_context& ctx) -> token_ptr
+{
+    if (ctx.current() != '\\') {return nullptr; }
+
+    if (ctx.peek() == 'b') {
+        static_cast<void>(ctx.next());
+        return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_check>)(cpp2::shared, "\\b", "word_boundary_token_matcher<char, false>"); 
+    }
+    else {if (ctx.peek() == 'B') {
+        static_cast<void>(ctx.next());
+        return CPP2_UFCS_TEMPLATE(cpp2_new<regex_token_check>)(cpp2::shared, "\\B", "word_boundary_token_matcher<char, true>"); 
+    }
+    else {
+        return nullptr; 
+    }}
+}
+
+#line 3508 "reflect.h2"
+    template <typename Error_out> regex_generator<Error_out>::regex_generator(cpp2::impl::in<std::string_view> r, Error_out const& e)
+        : regex{ r }
+        , error_out{ e }{
+
+#line 3511 "reflect.h2"
+    }
+
+#line 3513 "reflect.h2"
+    template <typename Error_out> [[nodiscard]] auto regex_generator<Error_out>::parse() & -> std::string
+    {
+        // Extract modifiers and adapt regex.
+        extract_modifiers();
+
+        parse_context parse_ctx {regex, error_out}; 
+        if (!(parse_ctx.parse(modifier))) {
+            return ""; 
+        }
+
+        source += "{\n";
+        source += " wrap: <Iter, CharT> type = {\n"; // TODO: Remove wrapper when template template parameters are available.
+        source += "  context: type == cpp2::regex::match_context<CharT, Iter, " + cpp2::to_string(parse_ctx.get_cur_group()) + ">;";
+
+        generation_context gen_ctx {}; 
+        source += gen_ctx.run(parse_ctx.get_as_token());
+        source += "  entry: (cur: Iter, inout ctx: context) -> cpp2::regex::match_return<Iter> = {\n";
+        source += "    ctx..set_group_start(0, cur);\n";
+        source += "    r := " + cpp2::to_string(gen_ctx.get_entry_func()) + "(cur, ctx, cpp2::regex::true_end_func());\n";
+        source += "    if r.matched { ctx..set_group_end(0, r.pos); }\n";
+        source += "    return r;\n";
+        source += "  }\n";
+
+        source += cpp2::move(gen_ctx).create_named_group_lookup(parse_ctx.named_groups);
+        source += "}\n";
+
+        auto string {(*cpp2::impl::assert_not_null(parse_ctx.get_as_token())).to_string()}; 
+        source += "  to_string: () -> std::string = { return R\"(" + cpp2::to_string(modifier_escape) + cpp2::to_string(cpp2::move(string)) + cpp2::to_string(modifier_escape) + cpp2::to_string(modifier) + ")\"; }\n";
+        source += "}\n";
+
+        static_cast<void>(cpp2::move(parse_ctx));
+
+        return source; 
+    }
+
+#line 3548 "reflect.h2"
+    template <typename Error_out> auto regex_generator<Error_out>::extract_modifiers() & -> void
+    {
+        if (regex.find_first_of("'/") == 0) {
+            char mod_token {CPP2_ASSERT_IN_BOUNDS_LITERAL(regex, 0)}; 
+
+            auto end_pos {regex.rfind(mod_token)}; 
+            if (end_pos != 0) {
+                // Found valid start end escape
+                modifier = regex.substr(end_pos + 1);
+                modifier_escape = cpp2::move(mod_token);
+                regex = regex.substr(1, cpp2::move(end_pos) - 1);
+            }
+        }
+    }
+
+#line 3564 "reflect.h2"
+template<typename Err> [[nodiscard]] auto generate_regex(cpp2::impl::in<std::string_view> regex, Err const& err) -> std::string
+{
+    regex_generator<Err> parser {regex, err}; 
+    auto r {parser.parse()}; 
+    static_cast<void>(cpp2::move(parser));
+    return r; 
+}
+
+}
+
+#line 3575 "reflect.h2"
+namespace meta {
+
+#line 3577 "reflect.h2"
 auto regex_gen(meta::type_declaration& t) -> void
 {
     auto has_default {false}; 
@@ -2216,7 +5237,7 @@ auto regex_gen(meta::type_declaration& t) -> void
     }
 }
 
-#line 1578 "reflect.h2"
+#line 3634 "reflect.h2"
 [[nodiscard]] auto apply_metafunctions(
     declaration_node& n, 
     type_declaration& rtype, 
@@ -2331,7 +5352,7 @@ auto regex_gen(meta::type_declaration& t) -> void
     return true; 
 }
 
-#line 1693 "reflect.h2"
+#line 3749 "reflect.h2"
 }
 
 }
